% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{assess_normality}
\alias{assess_normality}
\title{Assess normality of features}
\usage{
assess_normality(
  raw_data,
  excluded_columns,
  cpus = 1,
  out_prefix = "metapipe",
  plots_dir = getwd(),
  transf_vals = c(2, exp(1), 3, 4, 5, 6, 7, 8, 9, 10)
)
}
\arguments{
\item{raw_data}{data frame containing the raw data}

\item{excluded_columns}{vector containing the indices of the data set properties, excluded columns}

\item{cpus}{number of CPUS to be used}

\item{out_prefix}{prefix for output files, plots, etc.}

\item{plots_dir}{path to the directory where plots should be stored}

\item{transf_vals}{transformation values}
}
\value{
Structure containing the normalised data, if a suitable transformation was found
}
\description{
Assess normality of features
}
\examples{
example_data <- data.frame(ID = c(1,2,3,4,5), 
                           P1 = c("one", "two", "three", "four", "five"), 
                           T1 = rnorm(5), 
                           T2 = rnorm(5))
assess_normality(example_data, c(1, 2))

}
\seealso{
\code{\link{assess_normality_postprocessing}} and \code{\link{assess_normality_stats}}
}
