% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-assess-normality.R
\name{assess_normality_stats}
\alias{assess_normality_stats}
\title{Assess normality statistics}
\usage{
assess_normality_stats(out_prefix = "metapipe")
}
\arguments{
\item{out_prefix}{Prefix for output files and plots.}
}
\description{
Statistics for the normality assessment of the traits. Uses an output file 
generated by \code{\link{assess_normality_postprocessing}}, therefore, it 
cannot be executed without previously running the post-processing function.
}
\examples{
# Toy dataset
example_data <- data.frame(ID = c(1,2,3,4,5), 
                           P1 = c("one", "two", "three", "four", "five"), 
                           T1 = rnorm(5), 
                           T2 = rnorm(5))
example_data_normalised <- MetaPipe:::assess_normality_core(example_data, c(1, 2))
example_data_normalised_post <- 
 MetaPipe:::assess_normality_postprocessing(example_data, 
                                           c(1, 2), 
                                           example_data_normalised)
MetaPipe:::assess_normality_stats()


# F1 Seedling Ionomics dataset
data(ionomics) # Includes some missing data
ionomics_rev <- MetaPipe::replace_missing(ionomics, 
                                          excluded_columns = c(1, 2),
                                          replace_na =  TRUE)
ionomics_normalised <- 
  MetaPipe:::assess_normality_core(ionomics_rev,
                             excluded_columns = c(1, 2),
                             out_prefix = "ionomics",
                             transf_vals = c(2, exp(1)))
ionomics_normalised_post <- 
  MetaPipe:::assess_normality_postprocessing(ionomics_rev, 
                                            c(1, 2), 
                                            ionomics_normalised,
                                            out_prefix = "ionomics")
MetaPipe:::assess_normality_stats(out_prefix = "ionomics")

}
\seealso{
\code{\link{assess_normality_core}} and 
\code{\link{assess_normality_postprocessing}}
}
\keyword{internal}
