% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{check_alpha}
\alias{check_alpha}
\title{Check significance level}
\usage{
check_alpha(alpha)
}
\arguments{
\item{alpha}{Significance level.}
}
\value{
Nothing if alpha is valid, otherwise stop execution.
}
\description{
Check significance level (alpha) validity.
}
\examples{
MetaPipe:::check_alpha(0.5)
\dontrun{
    MetaPipe:::check_alpha(-0.5)
    MetaPipe:::check_alpha(1.5)
    MetaPipe:::check_alpha("1.5")
}

}
\keyword{internal}
