% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{check_transformation}
\alias{check_transformation}
\title{Check transformation}
\usage{
check_transformation(
  ref,
  new,
  transf,
  msg = " transformation does not normalise the data."
)
}
\arguments{
\item{ref}{Reference p-value.}

\item{new}{New p-value.}

\item{transf}{Transformation name (e.g. Log).}

\item{msg}{Message template.}
}
\value{
\code{TRUE} if the transformation normalised the data, \code{FALSE} 
otherwise.
}
\description{
Check whether or not the transformation normalises the data.
}
\examples{
MetaPipe:::check_transformation(0.05, 0.045, "Log")
MetaPipe:::check_transformation(0.05, 0.1, "Log")

}
\keyword{internal}
