% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_types}
\alias{check_types}
\title{Check variable types}
\usage{
check_types(raw_data, excluded_columns = NULL, numeric = TRUE, quiet = TRUE)
}
\arguments{
\item{raw_data}{data frame containing the raw data}

\item{excluded_columns}{list of excluded columns (won't be check)}

\item{numeric}{boolean to indicate the operation mode. If \code{TRUE} will
look up for non-numeric variables and exclude them, otherwise it will
filter numeric variables, including those with the following classes:
\code{complex} (e.g. \code{1+i})
\code{integer} (e.g. \code{1L})
\code{numeric} (e.g. \code{1}, \code{1.01})}

\item{quiet}{boolean to hide warning messages}
}
\value{
vector with the column indices that don't meet the checking criteria, 
only numeric (\code{numeric = TRUE}) or non-numeric (\code{numeric = FALSE}),  
if \code{excluded_columns != NULL}, returns a vector with both originally 
excluded and newly found column indices
}
\description{
Check variable types in a data frame. Useful whenever the data must be 
filtered prior some numerical analysis
}
\examples{
example_data <- data.frame(ID = c(1,2,3,4,5), 
                           P1 = c("one", "two", "three", "four", "five"), 
                           T1 = rnorm(5))
MetaPipe::check_types(example_data)
MetaPipe::check_types(example_data, numeric = FALSE)
MetaPipe::check_types(example_data, excluded_columns = 1, numeric = FALSE)
MetaPipe::check_types(example_data, quiet = FALSE)
}
