% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{compare_hist}
\alias{compare_hist}
\title{Compare histograms}
\usage{
compare_hist(original, transformed, trait, prefix, xlab)
}
\arguments{
\item{original}{Original data.}

\item{transformed}{Transformed data.}

\item{trait}{Trait name.}

\item{prefix}{File prefix.}

\item{xlab}{X-axis label.}
}
\description{
Compare two histograms from two different sources. For example original data 
and transformed data.
}
\examples{
norm_dist <- rnorm(100)
norm_dist_transformed <- norm_dist^2
MetaPipe::compare_hist(norm_dist, 
                       norm_dist_transformed, 
                       "XYZ", 
                       "xyz_hist", 
                       "x")
}
