% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lab_summaries.R
\name{create_lab_summaries}
\alias{create_lab_summaries}
\title{Creating Lab Summaries}
\usage{
create_lab_summaries(
  data,
  Project = NULL,
  Replication = NULL,
  Lab = NULL,
  DV = NULL,
  Group = NULL,
  IV = NULL,
  output_folder,
  suppress_list_output = FALSE
)
}
\arguments{
\item{data}{A list of data frames that contain the individual participant data.}

\item{Project}{The name of the columns in the list elements of "data" that contain the project name(s).}

\item{Replication}{The name of the columns in the list elements of "data" that contain the replication name(s).}

\item{Lab}{The name of the columns in the list elements of "data" that contain the replication names.}

\item{DV}{The name of the columns in the list elements of "data" that contain the (aggregated) dependent variable.}

\item{Group}{The name of the columns in the list elements of "data" that contain the (treatment/control) group identification. These should only contain values of 0 (control group), 1 (treatment group) and NA (unidentified).}

\item{output_folder}{Specify the output folder for the summaries and the codebook. If no folder is specified, the function will return its output only to the R environment (unless this is suppressed under suppress_list_output).}

\item{suppress_list_output}{Logical. FALSE by default. If FALSE, the function will return a list output to the environment, containing the lab summaries and the codebook. If TRUE, these are not returned to the environment.}
}
\value{
The function create_lab_summaries returns a list consisting of two elements: A codebook and a list of data frames. Each data frame contains all lab summary statistics for the according Replication/Effect.
The summary statistics returned (including their standard error) are the means and standard deviations for control and experimental groups, pooled standard deviations, raw mean differences and standardized mean differences (Hedge's g according to Borenstein).
}
\description{
Some description of the function
}
\examples{
## I don't know yet

}
