% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-qtl.R
\name{effect_plots}
\alias{effect_plots}
\title{Create effect plots}
\usage{
effect_plots(x_data_sim, qtl_data, cpus = 1, plots_dir = tempdir())
}
\arguments{
\item{x_data_sim}{Cross-data frame simulated with \code{qtl::sim.geno}.}

\item{qtl_data}{Significant QTL data.}

\item{cpus}{Number of CPUs to be used in the computation.}

\item{plots_dir}{Output directory for plots.}
}
\description{
Create effect plots for significant QTLs found with  
\code{\link{qtl_perm_test}}.
}
\seealso{
\code{\link{qtl_perm_test}}
}
