% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FatherRipariaMap}
\alias{FatherRipariaMap}
\title{Father Riparia Genetic Map}
\format{
A data frame with 296 rows and 1116 variables:
\describe{
    \item{ID}{Sample ID}
    \item{S1_61235}{Chromosome 1, position 61235}
    \item{S1_929902}{Chromosome 1, position 929902}
    \item{S1_666584}{Chromosome 1, position 666584}
    \item{S1_695458}{Chromosome 1, position 695458}
    \item{S1_666669}{Chromosome 1, position 666669}
    \item{S1_659894}{Chromosome 1, position 659894}
    \item{S1_1688964}{Chromosome 1, position 1688964}
    \item{S1_2537093}{Chromosome 1, position 2537093}
    \item{S1_2533272}{Chromosome 1, position 2533272}
    \item{S1_2533185}{Chromosome 1, position 2533185}
    \item{S1_2880124}{Chromosome 1, position 2880124}
    \item{S1_3062243}{Chromosome 1, position 3062243}
    \item{S1_3025844}{Chromosome 1, position 3025844}
    \item{S1_3466266}{Chromosome 1, position 3466266}
    \item{S1_3721305}{Chromosome 1, position 3721305}
    \item{S1_3757997}{Chromosome 1, position 3757997}
    \item{S1_3834552}{Chromosome 1, position 3834552}
    \item{S1_3889266}{Chromosome 1, position 3889266}
    \item{S1_4021184}{Chromosome 1, position 4021184}
    \item{S1_4096008}{Chromosome 1, position 4096008}
    \item{S1_4197004}{Chromosome 1, position 4197004}
    \item{S1_4279268}{Chromosome 1, position 4279268}
    \item{S1_4279373}{Chromosome 1, position 4279373}
    \item{S1_5767561}{Chromosome 1, position 5767561}
    \item{S1_6289531}{Chromosome 1, position 6289531}
    \item{S1_6289507}{Chromosome 1, position 6289507}
    \item{S1_6307621}{Chromosome 1, position 6307621}
    \item{S1_6691294}{Chromosome 1, position 6691294}
    \item{S1_6749776}{Chromosome 1, position 6749776}
    \item{S1_6743053}{Chromosome 1, position 6743053}
    \item{S1_6912064}{Chromosome 1, position 6912064}
    \item{S1_7042562}{Chromosome 1, position 7042562}
    \item{S1_7253240}{Chromosome 1, position 7253240}
    \item{S1_7356636}{Chromosome 1, position 7356636}
    \item{S1_7356638}{Chromosome 1, position 7356638}
    \item{S1_8516414}{Chromosome 1, position 8516414}
    \item{S1_9284785}{Chromosome 1, position 9284785}
    \item{S1_9424155}{Chromosome 1, position 9424155}
    \item{S1_9388557}{Chromosome 1, position 9388557}
    \item{S1_9060903}{Chromosome 1, position 9060903}
    \item{S1_9788875}{Chromosome 1, position 9788875}
    \item{S1_10830747}{Chromosome 1, position 10830747}
    \item{S1_11168597}{Chromosome 1, position 11168597}
    \item{S10_4018986}{Chromosome 10, position 4018986}
    \item{S1_13433555}{Chromosome 1, position 13433555}
    \item{S1_13411708}{Chromosome 1, position 13411708}
    \item{S17_4735662}{Chromosome 17, position 4735662}
    \item{S1_18216118}{Chromosome 1, position 18216118}
    \item{S1_18477299}{Chromosome 1, position 18477299}
    \item{S1_18957759}{Chromosome 1, position 18957759}
    \item{S1_19365070}{Chromosome 1, position 19365070}
    \item{S1_19637630}{Chromosome 1, position 19637630}
    \item{S1_19658204}{Chromosome 1, position 19658204}
    \item{S1_19817726}{Chromosome 1, position 19817726}
    \item{S1_19817735}{Chromosome 1, position 19817735}
    \item{S1_19817765}{Chromosome 1, position 19817765}
    \item{S1_20799803}{Chromosome 1, position 20799803}
    \item{S1_20799814}{Chromosome 1, position 20799814}
    \item{S1_21856314}{Chromosome 1, position 21856314}
    \item{S1_4912715}{Chromosome 1, position 4912715}
    \item{S1_23507587}{Chromosome 1, position 23507587}
    \item{S5_1411032}{Chromosome 5, position 1411032}
    \item{S1_24197159}{Chromosome 1, position 24197159}
    \item{S13_14783587}{Chromosome 13, position 14783587}
    \item{S2_681715}{Chromosome 2, position 681715}
    \item{S2_727595}{Chromosome 2, position 727595}
    \item{S2_684451}{Chromosome 2, position 684451}
    \item{S2_3454540}{Chromosome 2, position 3454540}
    \item{S2_3508231}{Chromosome 2, position 3508231}
    \item{S2_3835280}{Chromosome 2, position 3835280}
    \item{S2_4115561}{Chromosome 2, position 4115561}
    \item{S2_4350083}{Chromosome 2, position 4350083}
    \item{S2_4272908}{Chromosome 2, position 4272908}
    \item{S2_4272907}{Chromosome 2, position 4272907}
    \item{S2_4599939}{Chromosome 2, position 4599939}
    \item{S2_4650244}{Chromosome 2, position 4650244}
    \item{S2_4650229}{Chromosome 2, position 4650229}
    \item{S2_4650247}{Chromosome 2, position 4650247}
    \item{S2_6636380}{Chromosome 2, position 6636380}
    \item{S2_6967072}{Chromosome 2, position 6967072}
    \item{S2_7451312}{Chromosome 2, position 7451312}
    \item{S12_1266122}{Chromosome 12, position 1266122}
    \item{S2_8805548}{Chromosome 2, position 8805548}
    \item{S2_8581778}{Chromosome 2, position 8581778}
    \item{S2_8713638}{Chromosome 2, position 8713638}
    \item{S2_9166022}{Chromosome 2, position 9166022}
    \item{S7_3473511}{Chromosome 7, position 3473511}
    \item{S16_17497990}{Chromosome 16, position 17497990}
    \item{S2_14300760}{Chromosome 2, position 14300760}
    \item{S2_15380290}{Chromosome 2, position 15380290}
    \item{S2_15052944}{Chromosome 2, position 15052944}
    \item{S2_15428662}{Chromosome 2, position 15428662}
    \item{S2_15428758}{Chromosome 2, position 15428758}
    \item{S2_15380325}{Chromosome 2, position 15380325}
    \item{S2_15428718}{Chromosome 2, position 15428718}
    \item{S2_15052983}{Chromosome 2, position 15052983}
    \item{S2_15380310}{Chromosome 2, position 15380310}
    \item{S2_15380371}{Chromosome 2, position 15380371}
    \item{S2_17017117}{Chromosome 2, position 17017117}
    \item{S2_17019683}{Chromosome 2, position 17019683}
    \item{S2_17632093}{Chromosome 2, position 17632093}
    \item{S2_17714063}{Chromosome 2, position 17714063}
    \item{S2_17632126}{Chromosome 2, position 17632126}
    \item{S2_17631523}{Chromosome 2, position 17631523}
    \item{S2_17632108}{Chromosome 2, position 17632108}
    \item{S2_18456742}{Chromosome 2, position 18456742}
    \item{S2_18456340}{Chromosome 2, position 18456340}
    \item{S3_41385}{Chromosome 3, position 41385}
    \item{S3_534317}{Chromosome 3, position 534317}
    \item{S3_532958}{Chromosome 3, position 532958}
    \item{S3_532967}{Chromosome 3, position 532967}
    \item{S3_947527}{Chromosome 3, position 947527}
    \item{S3_969615}{Chromosome 3, position 969615}
    \item{S3_1001006}{Chromosome 3, position 1001006}
    \item{S3_1171786}{Chromosome 3, position 1171786}
    \item{S3_1407230}{Chromosome 3, position 1407230}
    \item{S3_1384415}{Chromosome 3, position 1384415}
    \item{S3_1484998}{Chromosome 3, position 1484998}
    \item{S3_1675026}{Chromosome 3, position 1675026}
    \item{S3_1992053}{Chromosome 3, position 1992053}
    \item{S3_1896075}{Chromosome 3, position 1896075}
    \item{S3_2224905}{Chromosome 3, position 2224905}
    \item{S3_2224904}{Chromosome 3, position 2224904}
    \item{S3_1992077}{Chromosome 3, position 1992077}
    \item{S3_1937647}{Chromosome 3, position 1937647}
    \item{S3_2458023}{Chromosome 3, position 2458023}
    \item{S3_2454850}{Chromosome 3, position 2454850}
    \item{S3_2439035}{Chromosome 3, position 2439035}
    \item{S3_2879059}{Chromosome 3, position 2879059}
    \item{S3_2876578}{Chromosome 3, position 2876578}
    \item{S3_2878868}{Chromosome 3, position 2878868}
    \item{S3_3151574}{Chromosome 3, position 3151574}
    \item{S3_3341202}{Chromosome 3, position 3341202}
    \item{S3_3396381}{Chromosome 3, position 3396381}
    \item{S3_3675428}{Chromosome 3, position 3675428}
    \item{S3_3675429}{Chromosome 3, position 3675429}
    \item{S3_4196171}{Chromosome 3, position 4196171}
    \item{S3_4228302}{Chromosome 3, position 4228302}
    \item{S3_4222506}{Chromosome 3, position 4222506}
    \item{S3_4520643}{Chromosome 3, position 4520643}
    \item{S3_4574460}{Chromosome 3, position 4574460}
    \item{S3_5278757}{Chromosome 3, position 5278757}
    \item{S3_5885144}{Chromosome 3, position 5885144}
    \item{S3_5884897}{Chromosome 3, position 5884897}
    \item{S3_6056253}{Chromosome 3, position 6056253}
    \item{S3_6843508}{Chromosome 3, position 6843508}
    \item{S3_6843554}{Chromosome 3, position 6843554}
    \item{S3_6843505}{Chromosome 3, position 6843505}
    \item{S3_6843565}{Chromosome 3, position 6843565}
    \item{S3_7104025}{Chromosome 3, position 7104025}
    \item{S3_7105462}{Chromosome 3, position 7105462}
    \item{S3_7193302}{Chromosome 3, position 7193302}
    \item{S3_7156370}{Chromosome 3, position 7156370}
    \item{S3_6406897}{Chromosome 3, position 6406897}
    \item{S3_9959283}{Chromosome 3, position 9959283}
    \item{S3_11937281}{Chromosome 3, position 11937281}
    \item{S3_11806166}{Chromosome 3, position 11806166}
    \item{S3_11746115}{Chromosome 3, position 11746115}
    \item{S5_4739215}{Chromosome 5, position 4739215}
    \item{S12_13636817}{Chromosome 12, position 13636817}
    \item{S3_19326615}{Chromosome 3, position 19326615}
    \item{S16_9895499}{Chromosome 16, position 9895499}
    \item{S16_9895496}{Chromosome 16, position 9895496}
    \item{S16_9895528}{Chromosome 16, position 9895528}
    \item{S4_453789}{Chromosome 4, position 453789}
    \item{S4_701098}{Chromosome 4, position 701098}
    \item{S4_700438}{Chromosome 4, position 700438}
    \item{S4_601440}{Chromosome 4, position 601440}
    \item{S4_977727}{Chromosome 4, position 977727}
    \item{S4_1283295}{Chromosome 4, position 1283295}
    \item{S4_1274066}{Chromosome 4, position 1274066}
    \item{S3_5683230}{Chromosome 3, position 5683230}
    \item{S4_2007592}{Chromosome 4, position 2007592}
    \item{S4_2501631}{Chromosome 4, position 2501631}
    \item{S4_2749888}{Chromosome 4, position 2749888}
    \item{S4_3479785}{Chromosome 4, position 3479785}
    \item{S4_3228910}{Chromosome 4, position 3228910}
    \item{S4_3207911}{Chromosome 4, position 3207911}
    \item{S4_3228959}{Chromosome 4, position 3228959}
    \item{S4_3902010}{Chromosome 4, position 3902010}
    \item{S4_3632159}{Chromosome 4, position 3632159}
    \item{S4_3632229}{Chromosome 4, position 3632229}
    \item{S4_6046549}{Chromosome 4, position 6046549}
    \item{S4_6381659}{Chromosome 4, position 6381659}
    \item{S4_6702831}{Chromosome 4, position 6702831}
    \item{S4_7806435}{Chromosome 4, position 7806435}
    \item{S4_7599187}{Chromosome 4, position 7599187}
    \item{S4_7599186}{Chromosome 4, position 7599186}
    \item{S9_22356376}{Chromosome 9, position 22356376}
    \item{S4_7664157}{Chromosome 4, position 7664157}
    \item{S4_8931079}{Chromosome 4, position 8931079}
    \item{S4_10577371}{Chromosome 4, position 10577371}
    \item{S4_9494990}{Chromosome 4, position 9494990}
    \item{S4_11890294}{Chromosome 4, position 11890294}
    \item{S4_12987946}{Chromosome 4, position 12987946}
    \item{S12_9215138}{Chromosome 12, position 9215138}
    \item{S18_29612869}{Chromosome 18, position 29612869}
    \item{S4_14520796}{Chromosome 4, position 14520796}
    \item{S17_16544364}{Chromosome 17, position 16544364}
    \item{S4_14829003}{Chromosome 4, position 14829003}
    \item{S4_15156973}{Chromosome 4, position 15156973}
    \item{S4_16242444}{Chromosome 4, position 16242444}
    \item{S4_16145098}{Chromosome 4, position 16145098}
    \item{S4_16592665}{Chromosome 4, position 16592665}
    \item{S4_16771616}{Chromosome 4, position 16771616}
    \item{S4_17801559}{Chromosome 4, position 17801559}
    \item{S4_17736774}{Chromosome 4, position 17736774}
    \item{S4_17966024}{Chromosome 4, position 17966024}
    \item{S4_18149111}{Chromosome 4, position 18149111}
    \item{S4_17941745}{Chromosome 4, position 17941745}
    \item{S4_19442593}{Chromosome 4, position 19442593}
    \item{S4_20127607}{Chromosome 4, position 20127607}
    \item{S4_20376448}{Chromosome 4, position 20376448}
    \item{S4_20476402}{Chromosome 4, position 20476402}
    \item{S4_20485683}{Chromosome 4, position 20485683}
    \item{S7_19909386}{Chromosome 7, position 19909386}
    \item{S4_20708423}{Chromosome 4, position 20708423}
    \item{S7_19909323}{Chromosome 7, position 19909323}
    \item{S4_21353666}{Chromosome 4, position 21353666}
    \item{S4_21609026}{Chromosome 4, position 21609026}
    \item{S4_21840191}{Chromosome 4, position 21840191}
    \item{S4_22856230}{Chromosome 4, position 22856230}
    \item{S4_23747765}{Chromosome 4, position 23747765}
    \item{S4_23684552}{Chromosome 4, position 23684552}
    \item{S4_23820188}{Chromosome 4, position 23820188}
    \item{S4_23839694}{Chromosome 4, position 23839694}
    \item{S4_24028683}{Chromosome 4, position 24028683}
    \item{S4_24561501}{Chromosome 4, position 24561501}
    \item{S5_735312}{Chromosome 5, position 735312}
    \item{S5_262303}{Chromosome 5, position 262303}
    \item{S5_1158453}{Chromosome 5, position 1158453}
    \item{S5_1158434}{Chromosome 5, position 1158434}
    \item{S5_1543083}{Chromosome 5, position 1543083}
    \item{S5_2724813}{Chromosome 5, position 2724813}
    \item{S5_3072237}{Chromosome 5, position 3072237}
    \item{S5_3309342}{Chromosome 5, position 3309342}
    \item{S5_4431511}{Chromosome 5, position 4431511}
    \item{S18_30393961}{Chromosome 18, position 30393961}
    \item{S5_5085070}{Chromosome 5, position 5085070}
    \item{S5_5561345}{Chromosome 5, position 5561345}
    \item{S5_5442598}{Chromosome 5, position 5442598}
    \item{S5_7077917}{Chromosome 5, position 7077917}
    \item{S5_7259503}{Chromosome 5, position 7259503}
    \item{S19_7239488}{Chromosome 19, position 7239488}
    \item{S5_8660631}{Chromosome 5, position 8660631}
    \item{S5_9112495}{Chromosome 5, position 9112495}
    \item{S5_9758363}{Chromosome 5, position 9758363}
    \item{S5_9694223}{Chromosome 5, position 9694223}
    \item{S5_9694146}{Chromosome 5, position 9694146}
    \item{S5_9538854}{Chromosome 5, position 9538854}
    \item{S4_2304111}{Chromosome 4, position 2304111}
    \item{S5_10114766}{Chromosome 5, position 10114766}
    \item{S5_12398332}{Chromosome 5, position 12398332}
    \item{S5_11607530}{Chromosome 5, position 11607530}
    \item{S5_11718096}{Chromosome 5, position 11718096}
    \item{S5_15323715}{Chromosome 5, position 15323715}
    \item{S5_15667054}{Chromosome 5, position 15667054}
    \item{S1_6307602}{Chromosome 1, position 6307602}
    \item{S5_18497601}{Chromosome 5, position 18497601}
    \item{S5_18497582}{Chromosome 5, position 18497582}
    \item{S5_20508317}{Chromosome 5, position 20508317}
    \item{S5_20469725}{Chromosome 5, position 20469725}
    \item{S5_20389519}{Chromosome 5, position 20389519}
    \item{S5_20142901}{Chromosome 5, position 20142901}
    \item{S5_20563021}{Chromosome 5, position 20563021}
    \item{S5_20877514}{Chromosome 5, position 20877514}
    \item{S5_20828642}{Chromosome 5, position 20828642}
    \item{S5_21366287}{Chromosome 5, position 21366287}
    \item{S5_21632590}{Chromosome 5, position 21632590}
    \item{S5_21729661}{Chromosome 5, position 21729661}
    \item{S5_22057342}{Chromosome 5, position 22057342}
    \item{S5_21614916}{Chromosome 5, position 21614916}
    \item{S5_22553215}{Chromosome 5, position 22553215}
    \item{S5_22996438}{Chromosome 5, position 22996438}
    \item{S5_23031628}{Chromosome 5, position 23031628}
    \item{S5_22996444}{Chromosome 5, position 22996444}
    \item{S5_22919343}{Chromosome 5, position 22919343}
    \item{S5_23031683}{Chromosome 5, position 23031683}
    \item{S5_23164911}{Chromosome 5, position 23164911}
    \item{S5_22996462}{Chromosome 5, position 22996462}
    \item{S5_22968587}{Chromosome 5, position 22968587}
    \item{S5_23871533}{Chromosome 5, position 23871533}
    \item{S5_24080763}{Chromosome 5, position 24080763}
    \item{S5_24031506}{Chromosome 5, position 24031506}
    \item{S5_24816875}{Chromosome 5, position 24816875}
    \item{S5_25196807}{Chromosome 5, position 25196807}
    \item{S5_25438857}{Chromosome 5, position 25438857}
    \item{S4_23787844}{Chromosome 4, position 23787844}
    \item{S6_395210}{Chromosome 6, position 395210}
    \item{S4_23787821}{Chromosome 4, position 23787821}
    \item{S6_780677}{Chromosome 6, position 780677}
    \item{S6_1563280}{Chromosome 6, position 1563280}
    \item{S6_1933490}{Chromosome 6, position 1933490}
    \item{S6_2124962}{Chromosome 6, position 2124962}
    \item{S6_2018054}{Chromosome 6, position 2018054}
    \item{S6_2338911}{Chromosome 6, position 2338911}
    \item{S6_2581999}{Chromosome 6, position 2581999}
    \item{S6_2530151}{Chromosome 6, position 2530151}
    \item{S6_3208601}{Chromosome 6, position 3208601}
    \item{S6_3524817}{Chromosome 6, position 3524817}
    \item{S6_3857152}{Chromosome 6, position 3857152}
    \item{S6_5018994}{Chromosome 6, position 5018994}
    \item{S6_5072530}{Chromosome 6, position 5072530}
    \item{S6_5065181}{Chromosome 6, position 5065181}
    \item{S6_5072505}{Chromosome 6, position 5072505}
    \item{S6_5391722}{Chromosome 6, position 5391722}
    \item{S6_6067964}{Chromosome 6, position 6067964}
    \item{S6_5878751}{Chromosome 6, position 5878751}
    \item{S6_5918482}{Chromosome 6, position 5918482}
    \item{S6_6037556}{Chromosome 6, position 6037556}
    \item{S6_6517464}{Chromosome 6, position 6517464}
    \item{S6_7240203}{Chromosome 6, position 7240203}
    \item{S6_7367026}{Chromosome 6, position 7367026}
    \item{S6_7422466}{Chromosome 6, position 7422466}
    \item{S6_7434965}{Chromosome 6, position 7434965}
    \item{S6_7434946}{Chromosome 6, position 7434946}
    \item{S6_8276089}{Chromosome 6, position 8276089}
    \item{S6_8177958}{Chromosome 6, position 8177958}
    \item{S6_11473551}{Chromosome 6, position 11473551}
    \item{S6_11398108}{Chromosome 6, position 11398108}
    \item{S9_4410620}{Chromosome 9, position 4410620}
    \item{S6_13928074}{Chromosome 6, position 13928074}
    \item{S6_14522304}{Chromosome 6, position 14522304}
    \item{S6_14522248}{Chromosome 6, position 14522248}
    \item{S6_14939919}{Chromosome 6, position 14939919}
    \item{S6_15724210}{Chromosome 6, position 15724210}
    \item{S6_15339772}{Chromosome 6, position 15339772}
    \item{S6_16115245}{Chromosome 6, position 16115245}
    \item{S6_18327662}{Chromosome 6, position 18327662}
    \item{S6_19048247}{Chromosome 6, position 19048247}
    \item{S6_19480467}{Chromosome 6, position 19480467}
    \item{S6_19923276}{Chromosome 6, position 19923276}
    \item{S6_20137833}{Chromosome 6, position 20137833}
    \item{S6_20191059}{Chromosome 6, position 20191059}
    \item{S6_20246431}{Chromosome 6, position 20246431}
    \item{S6_20445306}{Chromosome 6, position 20445306}
    \item{S6_20734091}{Chromosome 6, position 20734091}
    \item{S6_21133250}{Chromosome 6, position 21133250}
    \item{S6_22246199}{Chromosome 6, position 22246199}
    \item{S6_22246211}{Chromosome 6, position 22246211}
    \item{S7_1430462}{Chromosome 7, position 1430462}
    \item{S7_1430460}{Chromosome 7, position 1430460}
    \item{S7_1430459}{Chromosome 7, position 1430459}
    \item{S7_1033332}{Chromosome 7, position 1033332}
    \item{S7_915680}{Chromosome 7, position 915680}
    \item{S7_1235119}{Chromosome 7, position 1235119}
    \item{S7_1231538}{Chromosome 7, position 1231538}
    \item{S7_1227786}{Chromosome 7, position 1227786}
    \item{S7_1231413}{Chromosome 7, position 1231413}
    \item{S7_1455506}{Chromosome 7, position 1455506}
    \item{S7_1647439}{Chromosome 7, position 1647439}
    \item{S7_1750720}{Chromosome 7, position 1750720}
    \item{S7_1750716}{Chromosome 7, position 1750716}
    \item{S7_1770746}{Chromosome 7, position 1770746}
    \item{S7_2244947}{Chromosome 7, position 2244947}
    \item{S7_2206578}{Chromosome 7, position 2206578}
    \item{S7_2581409}{Chromosome 7, position 2581409}
    \item{S7_3843501}{Chromosome 7, position 3843501}
    \item{S7_4643469}{Chromosome 7, position 4643469}
    \item{S7_4652023}{Chromosome 7, position 4652023}
    \item{S7_5010810}{Chromosome 7, position 5010810}
    \item{S7_5266672}{Chromosome 7, position 5266672}
    \item{S7_5225574}{Chromosome 7, position 5225574}
    \item{S15_8375621}{Chromosome 15, position 8375621}
    \item{S7_6634754}{Chromosome 7, position 6634754}
    \item{S7_7022084}{Chromosome 7, position 7022084}
    \item{S7_7216324}{Chromosome 7, position 7216324}
    \item{S7_7583607}{Chromosome 7, position 7583607}
    \item{S7_8390900}{Chromosome 7, position 8390900}
    \item{S7_9383230}{Chromosome 7, position 9383230}
    \item{S7_9476959}{Chromosome 7, position 9476959}
    \item{S7_9476996}{Chromosome 7, position 9476996}
    \item{S7_10154065}{Chromosome 7, position 10154065}
    \item{S7_11512843}{Chromosome 7, position 11512843}
    \item{S12_16572791}{Chromosome 12, position 16572791}
    \item{S16_627321}{Chromosome 16, position 627321}
    \item{S5_15180614}{Chromosome 5, position 15180614}
    \item{S7_12136623}{Chromosome 7, position 12136623}
    \item{S7_12136635}{Chromosome 7, position 12136635}
    \item{S7_12162520}{Chromosome 7, position 12162520}
    \item{S14_13756234}{Chromosome 14, position 13756234}
    \item{S7_12552225}{Chromosome 7, position 12552225}
    \item{S5_12880208}{Chromosome 5, position 12880208}
    \item{S7_13101392}{Chromosome 7, position 13101392}
    \item{S14_15060535}{Chromosome 14, position 15060535}
    \item{S7_14622444}{Chromosome 7, position 14622444}
    \item{S7_14622513}{Chromosome 7, position 14622513}
    \item{S7_15270927}{Chromosome 7, position 15270927}
    \item{S7_15254134}{Chromosome 7, position 15254134}
    \item{S2_16557788}{Chromosome 2, position 16557788}
    \item{S18_18740935}{Chromosome 18, position 18740935}
    \item{S18_18740909}{Chromosome 18, position 18740909}
    \item{S18_18740961}{Chromosome 18, position 18740961}
    \item{S7_17610214}{Chromosome 7, position 17610214}
    \item{S7_17779698}{Chromosome 7, position 17779698}
    \item{S7_18079986}{Chromosome 7, position 18079986}
    \item{S7_18200112}{Chromosome 7, position 18200112}
    \item{S7_18172013}{Chromosome 7, position 18172013}
    \item{S7_18080011}{Chromosome 7, position 18080011}
    \item{S7_18080009}{Chromosome 7, position 18080009}
    \item{S7_18468799}{Chromosome 7, position 18468799}
    \item{S7_18468823}{Chromosome 7, position 18468823}
    \item{S7_18696664}{Chromosome 7, position 18696664}
    \item{S7_18819820}{Chromosome 7, position 18819820}
    \item{S7_18830228}{Chromosome 7, position 18830228}
    \item{S7_18819915}{Chromosome 7, position 18819915}
    \item{S7_20008924}{Chromosome 7, position 20008924}
    \item{S7_22156653}{Chromosome 7, position 22156653}
    \item{S7_22157803}{Chromosome 7, position 22157803}
    \item{S7_22192346}{Chromosome 7, position 22192346}
    \item{S7_22634499}{Chromosome 7, position 22634499}
    \item{S7_22735436}{Chromosome 7, position 22735436}
    \item{S7_22940661}{Chromosome 7, position 22940661}
    \item{S7_23044590}{Chromosome 7, position 23044590}
    \item{S7_23045098}{Chromosome 7, position 23045098}
    \item{S7_23125517}{Chromosome 7, position 23125517}
    \item{S7_23264421}{Chromosome 7, position 23264421}
    \item{S7_23660770}{Chromosome 7, position 23660770}
    \item{S7_23660739}{Chromosome 7, position 23660739}
    \item{S7_24696284}{Chromosome 7, position 24696284}
    \item{S7_24644008}{Chromosome 7, position 24644008}
    \item{S7_24892902}{Chromosome 7, position 24892902}
    \item{S7_25059843}{Chromosome 7, position 25059843}
    \item{S7_25500855}{Chromosome 7, position 25500855}
    \item{S7_25394583}{Chromosome 7, position 25394583}
    \item{S7_25459515}{Chromosome 7, position 25459515}
    \item{S7_25866846}{Chromosome 7, position 25866846}
    \item{S7_25946441}{Chromosome 7, position 25946441}
    \item{S7_26028904}{Chromosome 7, position 26028904}
    \item{S7_26783844}{Chromosome 7, position 26783844}
    \item{S7_27019046}{Chromosome 7, position 27019046}
    \item{S7_27015047}{Chromosome 7, position 27015047}
    \item{S1_5285762}{Chromosome 1, position 5285762}
    \item{S1_5285764}{Chromosome 1, position 5285764}
    \item{S8_237214}{Chromosome 8, position 237214}
    \item{S8_597446}{Chromosome 8, position 597446}
    \item{S8_2445085}{Chromosome 8, position 2445085}
    \item{S8_3002378}{Chromosome 8, position 3002378}
    \item{S16_1311790}{Chromosome 16, position 1311790}
    \item{S8_3764126}{Chromosome 8, position 3764126}
    \item{S8_4284131}{Chromosome 8, position 4284131}
    \item{S12_13707068}{Chromosome 12, position 13707068}
    \item{S8_5104468}{Chromosome 8, position 5104468}
    \item{S8_7219197}{Chromosome 8, position 7219197}
    \item{S8_7486223}{Chromosome 8, position 7486223}
    \item{S8_8735495}{Chromosome 8, position 8735495}
    \item{S8_8774657}{Chromosome 8, position 8774657}
    \item{S8_10222817}{Chromosome 8, position 10222817}
    \item{S8_10443307}{Chromosome 8, position 10443307}
    \item{S8_10699985}{Chromosome 8, position 10699985}
    \item{S8_10699975}{Chromosome 8, position 10699975}
    \item{S8_10993164}{Chromosome 8, position 10993164}
    \item{S8_11656031}{Chromosome 8, position 11656031}
    \item{S8_12528355}{Chromosome 8, position 12528355}
    \item{S8_12984202}{Chromosome 8, position 12984202}
    \item{S8_13494897}{Chromosome 8, position 13494897}
    \item{S8_13468856}{Chromosome 8, position 13468856}
    \item{S8_13784711}{Chromosome 8, position 13784711}
    \item{S8_13900785}{Chromosome 8, position 13900785}
    \item{S8_14133707}{Chromosome 8, position 14133707}
    \item{S8_14133706}{Chromosome 8, position 14133706}
    \item{S8_14202730}{Chromosome 8, position 14202730}
    \item{S8_14298254}{Chromosome 8, position 14298254}
    \item{S18_4298373}{Chromosome 18, position 4298373}
    \item{S8_15007978}{Chromosome 8, position 15007978}
    \item{S8_15007979}{Chromosome 8, position 15007979}
    \item{S8_16041762}{Chromosome 8, position 16041762}
    \item{S8_16155228}{Chromosome 8, position 16155228}
    \item{S8_17158629}{Chromosome 8, position 17158629}
    \item{S8_17158648}{Chromosome 8, position 17158648}
    \item{S8_17281173}{Chromosome 8, position 17281173}
    \item{S8_17272454}{Chromosome 8, position 17272454}
    \item{S8_17440416}{Chromosome 8, position 17440416}
    \item{S8_19508896}{Chromosome 8, position 19508896}
    \item{S8_19845299}{Chromosome 8, position 19845299}
    \item{S8_19820667}{Chromosome 8, position 19820667}
    \item{S8_21026494}{Chromosome 8, position 21026494}
    \item{S8_21195687}{Chromosome 8, position 21195687}
    \item{S8_21406499}{Chromosome 8, position 21406499}
    \item{S8_21905772}{Chromosome 8, position 21905772}
    \item{S8_21823877}{Chromosome 8, position 21823877}
    \item{S8_22140676}{Chromosome 8, position 22140676}
    \item{S9_185166}{Chromosome 9, position 185166}
    \item{S9_551908}{Chromosome 9, position 551908}
    \item{S9_742124}{Chromosome 9, position 742124}
    \item{S9_915820}{Chromosome 9, position 915820}
    \item{S9_1015693}{Chromosome 9, position 1015693}
    \item{S9_915836}{Chromosome 9, position 915836}
    \item{S9_1139404}{Chromosome 9, position 1139404}
    \item{S9_3871194}{Chromosome 9, position 3871194}
    \item{S9_3694806}{Chromosome 9, position 3694806}
    \item{S9_3468842}{Chromosome 9, position 3468842}
    \item{S9_3393247}{Chromosome 9, position 3393247}
    \item{S9_3468875}{Chromosome 9, position 3468875}
    \item{S9_3694878}{Chromosome 9, position 3694878}
    \item{S9_3817413}{Chromosome 9, position 3817413}
    \item{S9_4309236}{Chromosome 9, position 4309236}
    \item{S9_4635879}{Chromosome 9, position 4635879}
    \item{S9_4815598}{Chromosome 9, position 4815598}
    \item{S16_11684813}{Chromosome 16, position 11684813}
    \item{S9_5642741}{Chromosome 9, position 5642741}
    \item{S9_5734792}{Chromosome 9, position 5734792}
    \item{S9_5740700}{Chromosome 9, position 5740700}
    \item{S9_5740681}{Chromosome 9, position 5740681}
    \item{S9_5640646}{Chromosome 9, position 5640646}
    \item{S9_6353597}{Chromosome 9, position 6353597}
    \item{S14_11901912}{Chromosome 14, position 11901912}
    \item{S9_7547166}{Chromosome 9, position 7547166}
    \item{S9_8308769}{Chromosome 9, position 8308769}
    \item{S9_9217577}{Chromosome 9, position 9217577}
    \item{S9_9634401}{Chromosome 9, position 9634401}
    \item{S9_9634415}{Chromosome 9, position 9634415}
    \item{S13_15518595}{Chromosome 13, position 15518595}
    \item{S9_10400477}{Chromosome 9, position 10400477}
    \item{S9_10470454}{Chromosome 9, position 10470454}
    \item{S9_11038736}{Chromosome 9, position 11038736}
    \item{S9_9890565}{Chromosome 9, position 9890565}
    \item{S13_4173453}{Chromosome 13, position 4173453}
    \item{S9_21557614}{Chromosome 9, position 21557614}
    \item{S10_419927}{Chromosome 10, position 419927}
    \item{S10_371175}{Chromosome 10, position 371175}
    \item{S10_710185}{Chromosome 10, position 710185}
    \item{S10_1368821}{Chromosome 10, position 1368821}
    \item{S10_1362190}{Chromosome 10, position 1362190}
    \item{S10_1877004}{Chromosome 10, position 1877004}
    \item{S10_1876993}{Chromosome 10, position 1876993}
    \item{S10_1923183}{Chromosome 10, position 1923183}
    \item{S10_1923190}{Chromosome 10, position 1923190}
    \item{S10_2470297}{Chromosome 10, position 2470297}
    \item{S10_2868961}{Chromosome 10, position 2868961}
    \item{S10_3024940}{Chromosome 10, position 3024940}
    \item{S10_3959571}{Chromosome 10, position 3959571}
    \item{S10_6429561}{Chromosome 10, position 6429561}
    \item{S10_6502816}{Chromosome 10, position 6502816}
    \item{S10_7917650}{Chromosome 10, position 7917650}
    \item{S10_8058787}{Chromosome 10, position 8058787}
    \item{S10_8269070}{Chromosome 10, position 8269070}
    \item{S10_8447144}{Chromosome 10, position 8447144}
    \item{S10_8904936}{Chromosome 10, position 8904936}
    \item{S10_10919719}{Chromosome 10, position 10919719}
    \item{S10_12220756}{Chromosome 10, position 12220756}
    \item{S10_11441598}{Chromosome 10, position 11441598}
    \item{S10_12899710}{Chromosome 10, position 12899710}
    \item{S10_14083852}{Chromosome 10, position 14083852}
    \item{S10_14811103}{Chromosome 10, position 14811103}
    \item{S10_16614507}{Chromosome 10, position 16614507}
    \item{S10_21780436}{Chromosome 10, position 21780436}
    \item{S10_21960345}{Chromosome 10, position 21960345}
    \item{S10_22601505}{Chromosome 10, position 22601505}
    \item{S10_23152493}{Chromosome 10, position 23152493}
    \item{S10_23324430}{Chromosome 10, position 23324430}
    \item{S10_23478503}{Chromosome 10, position 23478503}
    \item{S11_1304562}{Chromosome 11, position 1304562}
    \item{S11_1894349}{Chromosome 11, position 1894349}
    \item{S11_1811646}{Chromosome 11, position 1811646}
    \item{S11_1973185}{Chromosome 11, position 1973185}
    \item{S11_1986925}{Chromosome 11, position 1986925}
    \item{S11_2056861}{Chromosome 11, position 2056861}
    \item{S11_2505433}{Chromosome 11, position 2505433}
    \item{S11_2567226}{Chromosome 11, position 2567226}
    \item{S11_2602688}{Chromosome 11, position 2602688}
    \item{S11_2886367}{Chromosome 11, position 2886367}
    \item{S11_3461670}{Chromosome 11, position 3461670}
    \item{S11_3968001}{Chromosome 11, position 3968001}
    \item{S11_5711212}{Chromosome 11, position 5711212}
    \item{S11_5371071}{Chromosome 11, position 5371071}
    \item{S11_5265819}{Chromosome 11, position 5265819}
    \item{S11_6152127}{Chromosome 11, position 6152127}
    \item{S11_5969557}{Chromosome 11, position 5969557}
    \item{S11_6152138}{Chromosome 11, position 6152138}
    \item{S11_7006742}{Chromosome 11, position 7006742}
    \item{S11_7260047}{Chromosome 11, position 7260047}
    \item{S11_8146822}{Chromosome 11, position 8146822}
    \item{S11_8146525}{Chromosome 11, position 8146525}
    \item{S11_10005713}{Chromosome 11, position 10005713}
    \item{S11_9838587}{Chromosome 11, position 9838587}
    \item{S11_10005714}{Chromosome 11, position 10005714}
    \item{S11_10139282}{Chromosome 11, position 10139282}
    \item{S18_29169830}{Chromosome 18, position 29169830}
    \item{S11_12185245}{Chromosome 11, position 12185245}
    \item{S11_12922083}{Chromosome 11, position 12922083}
    \item{S11_12922088}{Chromosome 11, position 12922088}
    \item{S11_15610317}{Chromosome 11, position 15610317}
    \item{S16_4387743}{Chromosome 16, position 4387743}
    \item{S11_16994463}{Chromosome 11, position 16994463}
    \item{S11_18410028}{Chromosome 11, position 18410028}
    \item{S11_18826413}{Chromosome 11, position 18826413}
    \item{S11_19405772}{Chromosome 11, position 19405772}
    \item{S11_19535876}{Chromosome 11, position 19535876}
    \item{S11_19614647}{Chromosome 11, position 19614647}
    \item{S11_19642437}{Chromosome 11, position 19642437}
    \item{S12_1266085}{Chromosome 12, position 1266085}
    \item{S12_2402996}{Chromosome 12, position 2402996}
    \item{S12_2468669}{Chromosome 12, position 2468669}
    \item{S12_2645369}{Chromosome 12, position 2645369}
    \item{S12_3272809}{Chromosome 12, position 3272809}
    \item{S12_3626640}{Chromosome 12, position 3626640}
    \item{S12_5321307}{Chromosome 12, position 5321307}
    \item{S12_4421166}{Chromosome 12, position 4421166}
    \item{S12_4460847}{Chromosome 12, position 4460847}
    \item{S12_4625727}{Chromosome 12, position 4625727}
    \item{S12_4985610}{Chromosome 12, position 4985610}
    \item{S12_4985627}{Chromosome 12, position 4985627}
    \item{S12_5287233}{Chromosome 12, position 5287233}
    \item{S12_5679188}{Chromosome 12, position 5679188}
    \item{S12_6209208}{Chromosome 12, position 6209208}
    \item{S12_6420693}{Chromosome 12, position 6420693}
    \item{S12_6469284}{Chromosome 12, position 6469284}
    \item{S12_6460434}{Chromosome 12, position 6460434}
    \item{S12_6521206}{Chromosome 12, position 6521206}
    \item{S12_6850379}{Chromosome 12, position 6850379}
    \item{S12_7330639}{Chromosome 12, position 7330639}
    \item{S12_14258968}{Chromosome 12, position 14258968}
    \item{S12_8761955}{Chromosome 12, position 8761955}
    \item{S12_8882507}{Chromosome 12, position 8882507}
    \item{S12_8983012}{Chromosome 12, position 8983012}
    \item{S12_9413160}{Chromosome 12, position 9413160}
    \item{S12_9775068}{Chromosome 12, position 9775068}
    \item{S12_9587137}{Chromosome 12, position 9587137}
    \item{S12_9663909}{Chromosome 12, position 9663909}
    \item{S12_12183577}{Chromosome 12, position 12183577}
    \item{S12_12152232}{Chromosome 12, position 12152232}
    \item{S12_12220260}{Chromosome 12, position 12220260}
    \item{S12_12183552}{Chromosome 12, position 12183552}
    \item{S12_12220276}{Chromosome 12, position 12220276}
    \item{S12_12974665}{Chromosome 12, position 12974665}
    \item{S12_12974887}{Chromosome 12, position 12974887}
    \item{S3_18984906}{Chromosome 3, position 18984906}
    \item{S12_14883785}{Chromosome 12, position 14883785}
    \item{S12_15653149}{Chromosome 12, position 15653149}
    \item{S12_15665349}{Chromosome 12, position 15665349}
    \item{S12_16418367}{Chromosome 12, position 16418367}
    \item{S12_17488444}{Chromosome 12, position 17488444}
    \item{S12_19118458}{Chromosome 12, position 19118458}
    \item{S12_19463132}{Chromosome 12, position 19463132}
    \item{S12_20072726}{Chromosome 12, position 20072726}
    \item{S12_20457660}{Chromosome 12, position 20457660}
    \item{S12_20782507}{Chromosome 12, position 20782507}
    \item{S12_20782482}{Chromosome 12, position 20782482}
    \item{S12_19075638}{Chromosome 12, position 19075638}
    \item{S3_17524411}{Chromosome 3, position 17524411}
    \item{S3_17524377}{Chromosome 3, position 17524377}
    \item{S12_20985597}{Chromosome 12, position 20985597}
    \item{S12_21068579}{Chromosome 12, position 21068579}
    \item{S12_21016691}{Chromosome 12, position 21016691}
    \item{S12_21424829}{Chromosome 12, position 21424829}
    \item{S12_21431223}{Chromosome 12, position 21431223}
    \item{S12_21619116}{Chromosome 12, position 21619116}
    \item{S12_21619135}{Chromosome 12, position 21619135}
    \item{S12_22588444}{Chromosome 12, position 22588444}
    \item{S15_1023023}{Chromosome 15, position 1023023}
    \item{S12_22597511}{Chromosome 12, position 22597511}
    \item{S12_23313501}{Chromosome 12, position 23313501}
    \item{S12_23545991}{Chromosome 12, position 23545991}
    \item{S12_23426060}{Chromosome 12, position 23426060}
    \item{S12_23430200}{Chromosome 12, position 23430200}
    \item{S12_23573521}{Chromosome 12, position 23573521}
    \item{S12_23756301}{Chromosome 12, position 23756301}
    \item{S12_23868695}{Chromosome 12, position 23868695}
    \item{S12_23853460}{Chromosome 12, position 23853460}
    \item{S12_23881923}{Chromosome 12, position 23881923}
    \item{S12_23881929}{Chromosome 12, position 23881929}
    \item{S13_492011}{Chromosome 13, position 492011}
    \item{S13_491967}{Chromosome 13, position 491967}
    \item{S13_869755}{Chromosome 13, position 869755}
    \item{S13_1189929}{Chromosome 13, position 1189929}
    \item{S13_1139928}{Chromosome 13, position 1139928}
    \item{S13_1305489}{Chromosome 13, position 1305489}
    \item{S13_1305382}{Chromosome 13, position 1305382}
    \item{S13_1911134}{Chromosome 13, position 1911134}
    \item{S13_1806912}{Chromosome 13, position 1806912}
    \item{S13_2107142}{Chromosome 13, position 2107142}
    \item{S13_2554407}{Chromosome 13, position 2554407}
    \item{S13_2272918}{Chromosome 13, position 2272918}
    \item{S13_2824036}{Chromosome 13, position 2824036}
    \item{S13_3078031}{Chromosome 13, position 3078031}
    \item{S13_3087282}{Chromosome 13, position 3087282}
    \item{S13_3646068}{Chromosome 13, position 3646068}
    \item{S13_3829146}{Chromosome 13, position 3829146}
    \item{S13_3829163}{Chromosome 13, position 3829163}
    \item{S13_3829207}{Chromosome 13, position 3829207}
    \item{S13_3829204}{Chromosome 13, position 3829204}
    \item{S13_4498185}{Chromosome 13, position 4498185}
    \item{S13_4705682}{Chromosome 13, position 4705682}
    \item{S13_4883318}{Chromosome 13, position 4883318}
    \item{S13_4932176}{Chromosome 13, position 4932176}
    \item{S13_5011753}{Chromosome 13, position 5011753}
    \item{S13_5011804}{Chromosome 13, position 5011804}
    \item{S13_5255336}{Chromosome 13, position 5255336}
    \item{S13_6653986}{Chromosome 13, position 6653986}
    \item{S13_6918755}{Chromosome 13, position 6918755}
    \item{S13_6918763}{Chromosome 13, position 6918763}
    \item{S13_8929490}{Chromosome 13, position 8929490}
    \item{S13_8971063}{Chromosome 13, position 8971063}
    \item{S13_8706967}{Chromosome 13, position 8706967}
    \item{S13_9506569}{Chromosome 13, position 9506569}
    \item{S13_9687839}{Chromosome 13, position 9687839}
    \item{S4_14674902}{Chromosome 4, position 14674902}
    \item{S13_10364271}{Chromosome 13, position 10364271}
    \item{S13_9922887}{Chromosome 13, position 9922887}
    \item{S13_9922847}{Chromosome 13, position 9922847}
    \item{S13_14516270}{Chromosome 13, position 14516270}
    \item{S9_10450473}{Chromosome 9, position 10450473}
    \item{S9_10450503}{Chromosome 9, position 10450503}
    \item{S13_14835143}{Chromosome 13, position 14835143}
    \item{S13_15188487}{Chromosome 13, position 15188487}
    \item{S13_15184596}{Chromosome 13, position 15184596}
    \item{S13_18428855}{Chromosome 13, position 18428855}
    \item{S13_18428911}{Chromosome 13, position 18428911}
    \item{S13_16718443}{Chromosome 13, position 16718443}
    \item{S13_16466363}{Chromosome 13, position 16466363}
    \item{S13_16846519}{Chromosome 13, position 16846519}
    \item{S13_18259226}{Chromosome 13, position 18259226}
    \item{S13_18364031}{Chromosome 13, position 18364031}
    \item{S13_18212211}{Chromosome 13, position 18212211}
    \item{S13_20851871}{Chromosome 13, position 20851871}
    \item{S13_20346527}{Chromosome 13, position 20346527}
    \item{S13_22230221}{Chromosome 13, position 22230221}
    \item{S13_22229184}{Chromosome 13, position 22229184}
    \item{S13_22492670}{Chromosome 13, position 22492670}
    \item{S13_22492639}{Chromosome 13, position 22492639}
    \item{S12_16438379}{Chromosome 12, position 16438379}
    \item{S13_23343299}{Chromosome 13, position 23343299}
    \item{S11_7993795}{Chromosome 11, position 7993795}
    \item{S13_25011963}{Chromosome 13, position 25011963}
    \item{S13_25837911}{Chromosome 13, position 25837911}
    \item{S13_26025336}{Chromosome 13, position 26025336}
    \item{S13_26071613}{Chromosome 13, position 26071613}
    \item{S13_26460953}{Chromosome 13, position 26460953}
    \item{S18_5527187}{Chromosome 18, position 5527187}
    \item{S13_26988205}{Chromosome 13, position 26988205}
    \item{S13_26226302}{Chromosome 13, position 26226302}
    \item{S13_27090277}{Chromosome 13, position 27090277}
    \item{S13_27456627}{Chromosome 13, position 27456627}
    \item{S13_27775437}{Chromosome 13, position 27775437}
    \item{S13_26762665}{Chromosome 13, position 26762665}
    \item{S13_28046237}{Chromosome 13, position 28046237}
    \item{S13_28328230}{Chromosome 13, position 28328230}
    \item{S13_28441366}{Chromosome 13, position 28441366}
    \item{S13_28874179}{Chromosome 13, position 28874179}
    \item{S13_29035074}{Chromosome 13, position 29035074}
    \item{S13_28629564}{Chromosome 13, position 28629564}
    \item{S14_130433}{Chromosome 14, position 130433}
    \item{S14_130435}{Chromosome 14, position 130435}
    \item{S14_155831}{Chromosome 14, position 155831}
    \item{S14_124731}{Chromosome 14, position 124731}
    \item{S14_227091}{Chromosome 14, position 227091}
    \item{S14_1276585}{Chromosome 14, position 1276585}
    \item{S14_1675170}{Chromosome 14, position 1675170}
    \item{S14_1644422}{Chromosome 14, position 1644422}
    \item{S14_1675247}{Chromosome 14, position 1675247}
    \item{S14_1771453}{Chromosome 14, position 1771453}
    \item{S10_7234127}{Chromosome 10, position 7234127}
    \item{S14_3153678}{Chromosome 14, position 3153678}
    \item{S14_3153680}{Chromosome 14, position 3153680}
    \item{S14_3160394}{Chromosome 14, position 3160394}
    \item{S14_3296208}{Chromosome 14, position 3296208}
    \item{S14_3322108}{Chromosome 14, position 3322108}
    \item{S14_3322054}{Chromosome 14, position 3322054}
    \item{S14_3596972}{Chromosome 14, position 3596972}
    \item{S14_4090452}{Chromosome 14, position 4090452}
    \item{S14_3891985}{Chromosome 14, position 3891985}
    \item{S14_5565801}{Chromosome 14, position 5565801}
    \item{S14_5694458}{Chromosome 14, position 5694458}
    \item{S14_5851353}{Chromosome 14, position 5851353}
    \item{S14_5737577}{Chromosome 14, position 5737577}
    \item{S14_5780286}{Chromosome 14, position 5780286}
    \item{S14_5771919}{Chromosome 14, position 5771919}
    \item{S14_6003433}{Chromosome 14, position 6003433}
    \item{S14_6783754}{Chromosome 14, position 6783754}
    \item{S14_6783728}{Chromosome 14, position 6783728}
    \item{S14_6919229}{Chromosome 14, position 6919229}
    \item{S14_7829515}{Chromosome 14, position 7829515}
    \item{S14_7776588}{Chromosome 14, position 7776588}
    \item{S14_8383937}{Chromosome 14, position 8383937}
    \item{S14_8660925}{Chromosome 14, position 8660925}
    \item{S14_9037195}{Chromosome 14, position 9037195}
    \item{S14_9198235}{Chromosome 14, position 9198235}
    \item{S14_10187131}{Chromosome 14, position 10187131}
    \item{S14_9198185}{Chromosome 14, position 9198185}
    \item{S14_11862177}{Chromosome 14, position 11862177}
    \item{S14_12700066}{Chromosome 14, position 12700066}
    \item{S14_14133313}{Chromosome 14, position 14133313}
    \item{S14_14133296}{Chromosome 14, position 14133296}
    \item{S4_10577387}{Chromosome 4, position 10577387}
    \item{S4_10577493}{Chromosome 4, position 10577493}
    \item{S14_16489468}{Chromosome 14, position 16489468}
    \item{S14_16487533}{Chromosome 14, position 16487533}
    \item{S14_17377883}{Chromosome 14, position 17377883}
    \item{S4_10577335}{Chromosome 4, position 10577335}
    \item{S14_20893124}{Chromosome 14, position 20893124}
    \item{S14_20102915}{Chromosome 14, position 20102915}
    \item{S14_20893101}{Chromosome 14, position 20893101}
    \item{S14_20893138}{Chromosome 14, position 20893138}
    \item{S14_20734690}{Chromosome 14, position 20734690}
    \item{S14_27233177}{Chromosome 14, position 27233177}
    \item{S14_21475133}{Chromosome 14, position 21475133}
    \item{S14_21475064}{Chromosome 14, position 21475064}
    \item{S14_22622919}{Chromosome 14, position 22622919}
    \item{S14_22579388}{Chromosome 14, position 22579388}
    \item{S18_32006717}{Chromosome 18, position 32006717}
    \item{S14_24001581}{Chromosome 14, position 24001581}
    \item{S14_24436637}{Chromosome 14, position 24436637}
    \item{S14_24674907}{Chromosome 14, position 24674907}
    \item{S14_24726221}{Chromosome 14, position 24726221}
    \item{S14_25060614}{Chromosome 14, position 25060614}
    \item{S14_24674005}{Chromosome 14, position 24674005}
    \item{S14_26697000}{Chromosome 14, position 26697000}
    \item{S14_27052967}{Chromosome 14, position 27052967}
    \item{S14_27100557}{Chromosome 14, position 27100557}
    \item{S14_27118850}{Chromosome 14, position 27118850}
    \item{S14_27233180}{Chromosome 14, position 27233180}
    \item{S14_27285728}{Chromosome 14, position 27285728}
    \item{S14_27565589}{Chromosome 14, position 27565589}
    \item{S14_27830765}{Chromosome 14, position 27830765}
    \item{S14_27931716}{Chromosome 14, position 27931716}
    \item{S14_28227292}{Chromosome 14, position 28227292}
    \item{S14_28537981}{Chromosome 14, position 28537981}
    \item{S14_28538005}{Chromosome 14, position 28538005}
    \item{S14_28560940}{Chromosome 14, position 28560940}
    \item{S14_28538025}{Chromosome 14, position 28538025}
    \item{S14_28925764}{Chromosome 14, position 28925764}
    \item{S14_28913313}{Chromosome 14, position 28913313}
    \item{S14_29535618}{Chromosome 14, position 29535618}
    \item{S14_29564069}{Chromosome 14, position 29564069}
    \item{S14_29768710}{Chromosome 14, position 29768710}
    \item{S14_29974825}{Chromosome 14, position 29974825}
    \item{S14_30257641}{Chromosome 14, position 30257641}
    \item{S2_16557224}{Chromosome 2, position 16557224}
    \item{S2_16557277}{Chromosome 2, position 16557277}
    \item{S11_11718739}{Chromosome 11, position 11718739}
    \item{S15_3214737}{Chromosome 15, position 3214737}
    \item{S4_23787790}{Chromosome 4, position 23787790}
    \item{S18_32006759}{Chromosome 18, position 32006759}
    \item{S18_18740943}{Chromosome 18, position 18740943}
    \item{S15_5592293}{Chromosome 15, position 5592293}
    \item{S15_6670011}{Chromosome 15, position 6670011}
    \item{S11_8037893}{Chromosome 11, position 8037893}
    \item{S1_1432206}{Chromosome 1, position 1432206}
    \item{S15_9620634}{Chromosome 15, position 9620634}
    \item{S17_17189706}{Chromosome 17, position 17189706}
    \item{S15_10104081}{Chromosome 15, position 10104081}
    \item{S15_10104107}{Chromosome 15, position 10104107}
    \item{S15_10966206}{Chromosome 15, position 10966206}
    \item{S15_11421787}{Chromosome 15, position 11421787}
    \item{S1_11209426}{Chromosome 1, position 11209426}
    \item{S1_11209437}{Chromosome 1, position 11209437}
    \item{S15_12062722}{Chromosome 15, position 12062722}
    \item{S15_12409905}{Chromosome 15, position 12409905}
    \item{S15_12409901}{Chromosome 15, position 12409901}
    \item{S15_12872594}{Chromosome 15, position 12872594}
    \item{S15_13258859}{Chromosome 15, position 13258859}
    \item{S15_14224942}{Chromosome 15, position 14224942}
    \item{S15_13824710}{Chromosome 15, position 13824710}
    \item{S15_13782184}{Chromosome 15, position 13782184}
    \item{S15_14318463}{Chromosome 15, position 14318463}
    \item{S15_14522792}{Chromosome 15, position 14522792}
    \item{S15_14940190}{Chromosome 15, position 14940190}
    \item{S15_15249593}{Chromosome 15, position 15249593}
    \item{S15_15640823}{Chromosome 15, position 15640823}
    \item{S15_15938964}{Chromosome 15, position 15938964}
    \item{S15_16026481}{Chromosome 15, position 16026481}
    \item{S15_16029228}{Chromosome 15, position 16029228}
    \item{S15_16775282}{Chromosome 15, position 16775282}
    \item{S15_16906382}{Chromosome 15, position 16906382}
    \item{S15_16944128}{Chromosome 15, position 16944128}
    \item{S15_18461648}{Chromosome 15, position 18461648}
    \item{S15_18442677}{Chromosome 15, position 18442677}
    \item{S15_18517441}{Chromosome 15, position 18517441}
    \item{S15_18608924}{Chromosome 15, position 18608924}
    \item{S15_18667076}{Chromosome 15, position 18667076}
    \item{S15_18935198}{Chromosome 15, position 18935198}
    \item{S15_18919732}{Chromosome 15, position 18919732}
    \item{S15_19084186}{Chromosome 15, position 19084186}
    \item{S15_19106930}{Chromosome 15, position 19106930}
    \item{S15_19449749}{Chromosome 15, position 19449749}
    \item{S4_6943168}{Chromosome 4, position 6943168}
    \item{S4_6943185}{Chromosome 4, position 6943185}
    \item{S15_19591533}{Chromosome 15, position 19591533}
    \item{S15_19735454}{Chromosome 15, position 19735454}
    \item{S15_19735459}{Chromosome 15, position 19735459}
    \item{S15_9531798}{Chromosome 15, position 9531798}
    \item{S16_157508}{Chromosome 16, position 157508}
    \item{S16_136028}{Chromosome 16, position 136028}
    \item{S16_535115}{Chromosome 16, position 535115}
    \item{S16_837377}{Chromosome 16, position 837377}
    \item{S16_1191190}{Chromosome 16, position 1191190}
    \item{S16_1261816}{Chromosome 16, position 1261816}
    \item{S16_1654149}{Chromosome 16, position 1654149}
    \item{S16_2778383}{Chromosome 16, position 2778383}
    \item{S16_2789979}{Chromosome 16, position 2789979}
    \item{S16_2848603}{Chromosome 16, position 2848603}
    \item{S16_2800127}{Chromosome 16, position 2800127}
    \item{S16_2996961}{Chromosome 16, position 2996961}
    \item{S16_3022275}{Chromosome 16, position 3022275}
    \item{S16_3022309}{Chromosome 16, position 3022309}
    \item{S12_11308887}{Chromosome 12, position 11308887}
    \item{S12_11308848}{Chromosome 12, position 11308848}
    \item{S16_6132045}{Chromosome 16, position 6132045}
    \item{S3_6059128}{Chromosome 3, position 6059128}
    \item{S16_6739476}{Chromosome 16, position 6739476}
    \item{S16_8000284}{Chromosome 16, position 8000284}
    \item{S3_6059194}{Chromosome 3, position 6059194}
    \item{S16_9192348}{Chromosome 16, position 9192348}
    \item{S16_11549139}{Chromosome 16, position 11549139}
    \item{S16_12991468}{Chromosome 16, position 12991468}
    \item{S16_13698008}{Chromosome 16, position 13698008}
    \item{S16_13698018}{Chromosome 16, position 13698018}
    \item{S16_14019789}{Chromosome 16, position 14019789}
    \item{S16_14019756}{Chromosome 16, position 14019756}
    \item{S13_3252820}{Chromosome 13, position 3252820}
    \item{S16_17498022}{Chromosome 16, position 17498022}
    \item{S16_17431826}{Chromosome 16, position 17431826}
    \item{S16_17458658}{Chromosome 16, position 17458658}
    \item{S16_18331100}{Chromosome 16, position 18331100}
    \item{S18_33757550}{Chromosome 18, position 33757550}
    \item{S16_19163249}{Chromosome 16, position 19163249}
    \item{S16_19218912}{Chromosome 16, position 19218912}
    \item{S16_19296557}{Chromosome 16, position 19296557}
    \item{S16_22011493}{Chromosome 16, position 22011493}
    \item{S16_22013957}{Chromosome 16, position 22013957}
    \item{S16_21189350}{Chromosome 16, position 21189350}
    \item{S16_20207555}{Chromosome 16, position 20207555}
    \item{S16_20734891}{Chromosome 16, position 20734891}
    \item{S16_20754235}{Chromosome 16, position 20754235}
    \item{S16_20556208}{Chromosome 16, position 20556208}
    \item{S16_10372933}{Chromosome 16, position 10372933}
    \item{S16_20280959}{Chromosome 16, position 20280959}
    \item{S16_22741961}{Chromosome 16, position 22741961}
    \item{S16_23326135}{Chromosome 16, position 23326135}
    \item{S16_23321614}{Chromosome 16, position 23321614}
    \item{S17_217269}{Chromosome 17, position 217269}
    \item{S17_427922}{Chromosome 17, position 427922}
    \item{S17_679913}{Chromosome 17, position 679913}
    \item{S17_1012250}{Chromosome 17, position 1012250}
    \item{S17_1012269}{Chromosome 17, position 1012269}
    \item{S17_1074151}{Chromosome 17, position 1074151}
    \item{S17_1251614}{Chromosome 17, position 1251614}
    \item{S17_1377163}{Chromosome 17, position 1377163}
    \item{S17_1485885}{Chromosome 17, position 1485885}
    \item{S17_1846189}{Chromosome 17, position 1846189}
    \item{S17_1846156}{Chromosome 17, position 1846156}
    \item{S17_3326406}{Chromosome 17, position 3326406}
    \item{S17_2902008}{Chromosome 17, position 2902008}
    \item{S17_2939211}{Chromosome 17, position 2939211}
    \item{S17_2933903}{Chromosome 17, position 2933903}
    \item{S17_3380646}{Chromosome 17, position 3380646}
    \item{S17_3492008}{Chromosome 17, position 3492008}
    \item{S17_3678280}{Chromosome 17, position 3678280}
    \item{S17_3925201}{Chromosome 17, position 3925201}
    \item{S17_4143051}{Chromosome 17, position 4143051}
    \item{S17_4200691}{Chromosome 17, position 4200691}
    \item{S17_4956722}{Chromosome 17, position 4956722}
    \item{S17_4983574}{Chromosome 17, position 4983574}
    \item{S4_19367472}{Chromosome 4, position 19367472}
    \item{S17_5533921}{Chromosome 17, position 5533921}
    \item{S17_5885473}{Chromosome 17, position 5885473}
    \item{S17_6004065}{Chromosome 17, position 6004065}
    \item{S17_6265079}{Chromosome 17, position 6265079}
    \item{S17_6265011}{Chromosome 17, position 6265011}
    \item{S17_6602678}{Chromosome 17, position 6602678}
    \item{S17_6744814}{Chromosome 17, position 6744814}
    \item{S17_10144665}{Chromosome 17, position 10144665}
    \item{S17_10144692}{Chromosome 17, position 10144692}
    \item{S17_10144600}{Chromosome 17, position 10144600}
    \item{S17_10615175}{Chromosome 17, position 10615175}
    \item{S17_11098255}{Chromosome 17, position 11098255}
    \item{S17_11143959}{Chromosome 17, position 11143959}
    \item{S17_11756483}{Chromosome 17, position 11756483}
    \item{S17_11804525}{Chromosome 17, position 11804525}
    \item{S17_11804177}{Chromosome 17, position 11804177}
    \item{S17_11804545}{Chromosome 17, position 11804545}
    \item{S15_13345468}{Chromosome 15, position 13345468}
    \item{S17_12124396}{Chromosome 17, position 12124396}
    \item{S17_12664468}{Chromosome 17, position 12664468}
    \item{S17_12418896}{Chromosome 17, position 12418896}
    \item{S17_12664542}{Chromosome 17, position 12664542}
    \item{S17_12823782}{Chromosome 17, position 12823782}
    \item{S17_12418905}{Chromosome 17, position 12418905}
    \item{S17_13760069}{Chromosome 17, position 13760069}
    \item{S4_13925699}{Chromosome 4, position 13925699}
    \item{S8_21089511}{Chromosome 8, position 21089511}
    \item{S17_18376592}{Chromosome 17, position 18376592}
    \item{S8_21089555}{Chromosome 8, position 21089555}
    \item{S18_2352415}{Chromosome 18, position 2352415}
    \item{S9_20274097}{Chromosome 9, position 20274097}
    \item{S18_3845009}{Chromosome 18, position 3845009}
    \item{S18_4382081}{Chromosome 18, position 4382081}
    \item{S18_4474754}{Chromosome 18, position 4474754}
    \item{S18_4493288}{Chromosome 18, position 4493288}
    \item{S18_4701680}{Chromosome 18, position 4701680}
    \item{S18_4997133}{Chromosome 18, position 4997133}
    \item{S18_5385362}{Chromosome 18, position 5385362}
    \item{S18_5834674}{Chromosome 18, position 5834674}
    \item{S18_6031168}{Chromosome 18, position 6031168}
    \item{S18_6796525}{Chromosome 18, position 6796525}
    \item{S18_6598486}{Chromosome 18, position 6598486}
    \item{S18_6938444}{Chromosome 18, position 6938444}
    \item{S18_7190789}{Chromosome 18, position 7190789}
    \item{S18_7329572}{Chromosome 18, position 7329572}
    \item{S18_7622618}{Chromosome 18, position 7622618}
    \item{S18_7943458}{Chromosome 18, position 7943458}
    \item{S18_8069708}{Chromosome 18, position 8069708}
    \item{S18_8085999}{Chromosome 18, position 8085999}
    \item{S18_8276990}{Chromosome 18, position 8276990}
    \item{S18_8536420}{Chromosome 18, position 8536420}
    \item{S18_8877775}{Chromosome 18, position 8877775}
    \item{S18_10262477}{Chromosome 18, position 10262477}
    \item{S18_10908692}{Chromosome 18, position 10908692}
    \item{S18_10905045}{Chromosome 18, position 10905045}
    \item{S18_11058645}{Chromosome 18, position 11058645}
    \item{S18_11016489}{Chromosome 18, position 11016489}
    \item{S18_11346823}{Chromosome 18, position 11346823}
    \item{S18_11355983}{Chromosome 18, position 11355983}
    \item{S18_12350424}{Chromosome 18, position 12350424}
    \item{S18_12137323}{Chromosome 18, position 12137323}
    \item{S18_12489391}{Chromosome 18, position 12489391}
    \item{S18_12438777}{Chromosome 18, position 12438777}
    \item{S18_12217024}{Chromosome 18, position 12217024}
    \item{S18_13727682}{Chromosome 18, position 13727682}
    \item{S18_13786991}{Chromosome 18, position 13786991}
    \item{S18_13705369}{Chromosome 18, position 13705369}
    \item{S18_14378796}{Chromosome 18, position 14378796}
    \item{S18_14371226}{Chromosome 18, position 14371226}
    \item{S18_14787346}{Chromosome 18, position 14787346}
    \item{S18_14787621}{Chromosome 18, position 14787621}
    \item{S18_19525649}{Chromosome 18, position 19525649}
    \item{S18_20287551}{Chromosome 18, position 20287551}
    \item{S18_20330236}{Chromosome 18, position 20330236}
    \item{S17_4735661}{Chromosome 17, position 4735661}
    \item{S18_23240627}{Chromosome 18, position 23240627}
    \item{S18_23946070}{Chromosome 18, position 23946070}
    \item{S18_26558746}{Chromosome 18, position 26558746}
    \item{S18_26379007}{Chromosome 18, position 26379007}
    \item{S18_26379010}{Chromosome 18, position 26379010}
    \item{S16_8638511}{Chromosome 16, position 8638511}
    \item{S7_17086479}{Chromosome 7, position 17086479}
    \item{S18_29975041}{Chromosome 18, position 29975041}
    \item{S18_30764626}{Chromosome 18, position 30764626}
    \item{S18_30764661}{Chromosome 18, position 30764661}
    \item{S5_4316057}{Chromosome 5, position 4316057}
    \item{S18_30764615}{Chromosome 18, position 30764615}
    \item{S18_30764638}{Chromosome 18, position 30764638}
    \item{S18_30764614}{Chromosome 18, position 30764614}
    \item{S18_31160265}{Chromosome 18, position 31160265}
    \item{S18_31160355}{Chromosome 18, position 31160355}
    \item{S18_31643840}{Chromosome 18, position 31643840}
    \item{S18_31624774}{Chromosome 18, position 31624774}
    \item{S18_31691725}{Chromosome 18, position 31691725}
    \item{S18_31822254}{Chromosome 18, position 31822254}
    \item{S18_33376662}{Chromosome 18, position 33376662}
    \item{S18_33123281}{Chromosome 18, position 33123281}
    \item{S18_31952418}{Chromosome 18, position 31952418}
    \item{S18_32516041}{Chromosome 18, position 32516041}
    \item{S18_32516030}{Chromosome 18, position 32516030}
    \item{S18_34125332}{Chromosome 18, position 34125332}
    \item{S18_32027400}{Chromosome 18, position 32027400}
    \item{S18_32516169}{Chromosome 18, position 32516169}
    \item{S19_472225}{Chromosome 19, position 472225}
    \item{S19_867297}{Chromosome 19, position 867297}
    \item{S19_1680257}{Chromosome 19, position 1680257}
    \item{S19_1896354}{Chromosome 19, position 1896354}
    \item{S19_1931387}{Chromosome 19, position 1931387}
    \item{S19_2154081}{Chromosome 19, position 2154081}
    \item{S19_2154118}{Chromosome 19, position 2154118}
    \item{S19_2291878}{Chromosome 19, position 2291878}
    \item{S19_3090390}{Chromosome 19, position 3090390}
    \item{S10_4125712}{Chromosome 10, position 4125712}
    \item{S19_3474455}{Chromosome 19, position 3474455}
    \item{S19_4566708}{Chromosome 19, position 4566708}
    \item{S19_4165852}{Chromosome 19, position 4165852}
    \item{S19_4808973}{Chromosome 19, position 4808973}
    \item{S19_5427703}{Chromosome 19, position 5427703}
    \item{S19_5403377}{Chromosome 19, position 5403377}
    \item{S19_5427686}{Chromosome 19, position 5427686}
    \item{S19_5427685}{Chromosome 19, position 5427685}
    \item{S19_5994599}{Chromosome 19, position 5994599}
    \item{S19_5880275}{Chromosome 19, position 5880275}
    \item{S19_6156678}{Chromosome 19, position 6156678}
    \item{S19_6186424}{Chromosome 19, position 6186424}
    \item{S19_6173411}{Chromosome 19, position 6173411}
    \item{S19_6333674}{Chromosome 19, position 6333674}
    \item{S19_6348968}{Chromosome 19, position 6348968}
    \item{S19_6552977}{Chromosome 19, position 6552977}
    \item{S19_6629721}{Chromosome 19, position 6629721}
    \item{S19_6695063}{Chromosome 19, position 6695063}
    \item{S16_1311752}{Chromosome 16, position 1311752}
    \item{S19_6695026}{Chromosome 19, position 6695026}
    \item{S19_6877038}{Chromosome 19, position 6877038}
    \item{S19_6860415}{Chromosome 19, position 6860415}
    \item{S19_7115370}{Chromosome 19, position 7115370}
    \item{S19_7443826}{Chromosome 19, position 7443826}
    \item{S19_7710108}{Chromosome 19, position 7710108}
    \item{S19_12204443}{Chromosome 19, position 12204443}
    \item{S19_17018544}{Chromosome 19, position 17018544}
    \item{S18_29612881}{Chromosome 18, position 29612881}
    \item{S19_16872865}{Chromosome 19, position 16872865}
    \item{S19_14740600}{Chromosome 19, position 14740600}
    \item{S19_18306424}{Chromosome 19, position 18306424}
    \item{S11_18407277}{Chromosome 11, position 18407277}
    \item{S17_12809119}{Chromosome 17, position 12809119}
    \item{S18_33759666}{Chromosome 18, position 33759666}
    \item{S19_20186412}{Chromosome 19, position 20186412}
    \item{S19_20536066}{Chromosome 19, position 20536066}
    \item{S19_21678572}{Chromosome 19, position 21678572}
    \item{S19_21912909}{Chromosome 19, position 21912909}
    \item{S4_9495191}{Chromosome 4, position 9495191}
    \item{S19_22763519}{Chromosome 19, position 22763519}
    \item{S19_22635646}{Chromosome 19, position 22635646}
    \item{S19_22635698}{Chromosome 19, position 22635698}
    \item{S19_22968690}{Chromosome 19, position 22968690}
    \item{S19_23054928}{Chromosome 19, position 23054928}
    \item{S19_23343775}{Chromosome 19, position 23343775}
    \item{S19_21023262}{Chromosome 19, position 21023262}
    \item{S19_24066053}{Chromosome 19, position 24066053}
}
}
\source{
\url{https://fennell-lab.netlify.app}
}
\usage{
data(FatherRipariaMap)
}
\description{
A dataset containing information of genotypes within each chromosome.
}
\author{
Gaurab Bhattarai \email{gbhattarai@uga.edu}
}
\keyword{datasets}
