% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_pipeline.R
\name{full_pipeline}
\alias{full_pipeline}
\title{Full Pipeline Function}
\usage{
full_pipeline(
  Data,
  Project = NULL,
  Replication = NULL,
  Lab = NULL,
  DV = NULL,
  Group = NULL,
  IV = NULL,
  output_folder,
  folder_name = NULL,
  suppress_list_output = FALSE
)
}
\description{
Full Pipeline Function
}
