% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{generate_hist}
\alias{generate_hist}
\title{This function generates a single histogram}
\usage{
generate_hist(data, feature, prefix = "metapipe", xlab = NULL)
}
\arguments{
\item{data}{Histogram data}

\item{feature}{Feature name}

\item{prefix}{File prefix [default: metapipe]}

\item{xlab}{x-axis label [default: NULL]}
}
\description{
This function generates a single histogram
}
\examples{
norm_dist <- rnorm(100)
generate_hist(norm_dist, "XYZ", "xyz_hist", "x")
}
