% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{ggplot_save}
\alias{ggplot_save}
\title{Save \code{gpplot2} plot as PNG}
\usage{
ggplot_save(plt_obj, name, width = 6, height = 6)
}
\arguments{
\item{plt_obj}{Graphical object.}

\item{name}{Output name with or without path.}

\item{width}{Width in inches.}

\item{height}{Height in inches.}
}
\description{
Save a graphical object generated with \code{ggplot2} to disk in 
Portable Network Graphics (PNG) format.
}
\examples{
plt_obj <- ggplot2::qplot(rnorm(100))
MetaPipe:::ggplot_save(plt_obj, "hist")

# Clean up example outputs
MetaPipe:::tidy_up("hist")

}
\keyword{internal}
