% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{hex_logo}
\alias{hex_logo}
\title{Hexagonal logo}
\usage{
hex_logo(
  subplot = system.file("images/lab-2.png", package = "MetaPipe"),
  dpi = 800,
  h_color = "#000000",
  h_fill = "#363b74",
  output = system.file("images/metapipe.png", package = "MetaPipe"),
  p_color = "#eeeeee"
)
}
\arguments{
\item{subplot}{Subplot/image/icon to be used in the background.}

\item{dpi}{Dots per inch.}

\item{h_color}{Hexagon edge colour.}

\item{h_fill}{Hexagon filling colour.}

\item{output}{Output filename and path.}

\item{p_color}{Package name colour.}
}
\description{
Create Hexagonal logo for MetaPipe
}
\examples{
MetaPipe:::hex_logo(output = "hex-logo.png")

}
\keyword{internal}
