% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl-utils.R
\name{is_pseudo_marker}
\alias{is_pseudo_marker}
\title{Verifies whether or not a string is a pseudo-marker, contains the substring
    \code{'loc'}}
\usage{
is_pseudo_marker(marker)
}
\arguments{
\item{marker}{string with the marker}
}
\value{
TRUE or FALSE
}
\description{
Verifies whether or not a string is a pseudo-marker, contains the substring
    \code{'loc'}
}
\examples{
is_pseudo_marker('c1.loc1')
is_pseudo_marker('S1_2345')
}
