% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-qtl.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(input, wdir = here::here(), contents = "raw", ...)
}
\arguments{
\item{input}{Input data. It can be a string with the filename or a 
data frame.}

\item{wdir}{Working directory.}

\item{contents}{Contents of \code{input}.}

\item{...}{
  Arguments passed on to \code{\link[utils:read.table]{utils::read.csv}}
  \describe{
    \item{\code{header}}{a logical value indicating whether the file contains the
    names of the variables as its first line.  If missing, the value is
    determined from the file format: \code{header} is set to \code{TRUE}
    if and only if the first row contains one fewer field than the
    number of columns.}
    \item{\code{sep}}{the field separator character.  Values on each line of the
    file are separated by this character.  If \code{sep = ""} (the
    default for \code{read.table}) the separator is \sQuote{white space},
    that is one or more spaces, tabs, newlines or carriage returns.}
    \item{\code{quote}}{the set of quoting characters. To disable quoting
    altogether, use \code{quote = ""}.  See \code{\link{scan}} for the
    behaviour on quotes embedded in quotes.  Quoting is only considered
    for columns read as character, which is all of them unless
    \code{colClasses} is specified.}
    \item{\code{dec}}{the character used in the file for decimal points.}
    \item{\code{fill}}{logical. If \code{TRUE} then in case the rows have unequal
    length, blank fields are implicitly added.  See \sQuote{Details}.}
    \item{\code{comment.char}}{character: a character vector of length one
    containing a single character or an empty string.  Use \code{""} to
    turn off the interpretation of comments altogether.}
  }}
}
\value{
Data frame.
}
\description{
Load data from a CSV file or verify that input is a data frame.
}
\keyword{internal}
