% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{load_raw}
\alias{load_raw}
\title{Load raw data}
\usage{
load_raw(raw_data_filename, excluded_columns = NULL)
}
\arguments{
\item{raw_data_filename}{filename containing the raw data, it can be a
relative path or an absolute path.}

\item{excluded_columns}{vector containing the indices of the dataset 
properties that are non-numeric, excluded columns}
}
\value{
data frame with the pre-processed raw data
}
\description{
Load raw data from disk and aggregates using the \code{mean} function 
observations with duplicated IDs (first column). Non-numeric columns must
be excluded using the \code{excluded_columns} parameter.
}
\examples{
ionomics_path <- system.file("extdata", 
                             "ionomics.csv", 
                             package = "MetaPipe", 
                             mustWork = TRUE)
ionomics <- MetaPipe::load_raw(ionomics_path)
}
