% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{log_transformation}
\alias{log_transformation}
\title{Logarithmic transformation}
\usage{
log_transformation(
  data,
  trait = "DATA",
  alpha = 0.05,
  transf_vals = c(2, exp(1), 3, 4, 5, 6, 7, 8, 9, 10),
  plots_prefix = "HIST",
  digits = 6
)
}
\arguments{
\item{data}{Original data.}

\item{trait}{Trait name.}

\item{alpha}{Significance level.}

\item{transf_vals}{Transformation values.}

\item{plots_prefix}{Prefix for plots with or without path.}

\item{digits}{Significant digits to compare p-values of transformations.}
}
\value{
Data structure containing the normalised data, \code{NULL} if no 
transformation was performed.
}
\description{
Normalise data with a log transformation.
}
\examples{
set.seed(123)
data <- rnorm(100, 5)
MetaPipe::log_transformation(2 ^ data, "EXP_2")

}
\seealso{
\code{\link{power_transformation}} and
\code{\link{root_transformation}}
}
