% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_lab_summaries.R
\name{merge_lab_summaries}
\alias{merge_lab_summaries}
\title{Merging Lab Summaries}
\usage{
merge_lab_summaries(data, output_folder, suppress_list_output = FALSE)
}
\arguments{
\item{data}{Either a list object or a path to a folder containing the lab summaries.}

\item{output_folder}{define a path to which the merged lab summaries and the codebook are exported.}

\item{suppress_list_output}{a logical indicating whether results should be returned in R.}
}
\value{
If all .csv files or list elements from the data input of the function are named coherently, the function returns a list with two elements (merged df & codebook) and optionally a folder (in case output_path is specified)
}
\description{
Some description of the function
}
\examples{
## Example with a list object (number of columns drastically reduced for illustration)
test_list <- list(data.frame(Project = c(1,1), Replication = c(1,1), Lab = c(1,2), T_M = c(0.5, 0.7)),  data.frame(Project = c(1,1), Replication = c(2,2), Lab = c(1,2), T_M = c(3, 5)))
merge_lab_summaries(data = test_list)
# returns a list with the "merged lab summaries" and the "codebook"
}
