% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\docType{package}
\name{MetaPipe-package}
\alias{MetaPipe}
\alias{MetaPipe-package}
\title{MetaPipe: A High-Performance Computing Pipeline for QTL Mapping of Large Ionomic and Metabolomic Datasets}
\description{
Pipeline for Quantitative Trait Locus (QTL) mapping of large ionomic and metabolomic datasets. It uses a High-Performance Computing (HPC) approach to speed up the computations performed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/villegar/MetaPipe/}
  \item Report bugs at \url{https://github.com/villegar/MetaPipe/issues/}
}

}
\author{
\strong{Maintainer}: Roberto Villegas-Diaz \email{villegas.roberto@hotmail.com} (\href{https://orcid.org/0000-0001-5036-8661}{ORCID})

Authors:
\itemize{
  \item Dilmini Alahakoon \email{dilmini.alahakoon@sdstate.edu}
  \item Anne Fennell \email{anne.fennell@sdstate.edu} (\href{https://orcid.org/0000-0002-4234-6419}{ORCID}) [thesis advisor]
}

Other contributors:
\itemize{
  \item Jason P. Londo \email{jason.londo@usda.gov} (\href{https://orcid.org/0000-0003-0535-4820}{ORCID}) [data contributor]
  \item Gaurab Bhattarai \email{gbhattarai@uga.edu} (\href{https://orcid.org/0000-0002-2899-7725}{ORCID}) [data contributor]
  \item Laszlo G. Kovacs \email{laszlokovacs@missouristate.edu} [data contributor]
}

}
\keyword{internal}
