% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{paretoscale}
\alias{paretoscale}
\title{Pareto scaling}
\usage{
paretoscale(z)
}
\arguments{
\item{z}{Original 2-dimensional data.}
}
\value{
Scaled data.
}
\description{
Perform a pareto scaling.
}
\examples{
set.seed(123)
data <- matrix(rnorm(100, 5), ncol = 2)
data_new <- MetaPipe:::paretoscale(data)

}
\references{
Grace, Stephen & Hudson, Dane. (2016). Processing and Visualization of 
Metabolomics Data Using R. \doi{10.5772/65405}.
}
\keyword{internal}
