% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{power_transformation}
\alias{power_transformation}
\title{Normalise data with a power transformation}
\usage{
power_transformation(
  data,
  trait = "DATA",
  alpha = 0.05,
  transf = c(2, exp(1), 3, 4, 5, 6, 7, 8, 9, 10),
  plots_prefix = "HIST",
  digits = 6
)
}
\arguments{
\item{data}{original data}

\item{trait}{trait name}

\item{alpha}{significance level}

\item{transf}{transformation values}

\item{plots_prefix}{prefix for plots with or without path}

\item{digits}{significant digits to compare p-values of transformations}
}
\value{
data structure containing the normalised data, NULL if no 
transformation was performed
}
\description{
Normalise data with a power transformation
}
\examples{
\dontrun{
    set.seed(123)
    data <- rnorm(100, 5)
    power_transformation(sqrt(data), "ROOT_2")
}
}
