% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{qtl_perm_test}
\alias{qtl_perm_test}
\title{Perform QTL mapping permutation test using scanone to obtain significant QTLs}
\usage{
qtl_perm_test(
  x_data,
  cpus = 1,
  qtl_method = "scanone",
  raw_data_normalised = NULL,
  lod_threshold = 3,
  parametric = TRUE,
  n_perm = 1000,
  plots_dir = getwd(),
  ...
)
}
\arguments{
\item{x_data}{cross-data containing genetic map data and features}

\item{cpus}{number of CPUS to be used}

\item{qtl_method}{QTL mapping method [default: scanone]}

\item{raw_data_normalised}{normalised raw data}

\item{lod_threshold}{LOD score threshold to look up for significant QTLs}

\item{parametric}{boolean flag to indicate whether or not x_data is parametric}

\item{n_perm}{number of permutations}

\item{plots_dir}{}

\item{...}{S4 parameters for R/qtl library}
}
\value{

}
\description{
Perform QTL mapping permutation test using scanone to obtain significant QTLs
}
\examples{

}
\seealso{
\code{\link{qtl_scone}}
}
