% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{qtl_scone}
\alias{qtl_scone}
\title{Perform QTL mapping}
\usage{
qtl_scone(x_data, cpus = 1, ...)
}
\arguments{
\item{x_data}{Cross-data frame containing genetic map data and traits.}

\item{cpus}{Number of CPUs to be used in the computation.}

\item{...}{Optional parameters for 
\code{\link[qtl:scanone]{qtl:scanone(...)}}.}
}
\value{
Data frame containing the LOD scores.
}
\description{
Perform QTL mapping using the \code{\link[qtl:scanone]{qtl:scanone(...)}} 
function to obtain LOD scores for all traits, peak positions, and markers.
}
\examples{
# Create toy dataset
excluded_columns <- c(1, 2)
population <- 5
seed <- 1
example_data <- data.frame(ID = 1:population,
                           P1 = c("one", "two", "three", "four", "five"),
                           T1 = rnorm(population),
                           T2 = rnorm(population))
                           
output <- MetaPipe::assess_normality(example_data, 
                                     excluded_columns,
                                     out_prefix = here::here("metapipe"),
                                     show_stats = FALSE)

# Create and store random genetic map [for testing only]
genetic_map <- MetaPipe:::random_map(population = population, 
                                     seed = seed)
write.csv(genetic_map, 
          here::here("metapipe_genetic_map.csv"), 
          row.names = FALSE)

# Load cross file with genetic map and raw data for normal traits
x <- qtl::read.cross(format = "csvs", 
                     dir = here::here(),
                     genfile = "metapipe_genetic_map.csv",
                     phefile = "metapipe_raw_data_norm.csv")
set.seed(seed)
x <- qtl::jittermap(x)
x <- qtl::calc.genoprob(x, step = 1, error.prob = 0.001)
x_scone <- qtl_scone(x, 1, model = "normal", method = "hk")

}
\seealso{
Other QTL mapping functions: 
\code{\link{qtl_perm_test}()}
}
\concept{QTL mapping functions}
