% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{random_genotypes}
\alias{random_genotypes}
\title{Generate a pseudo-random genotypes}
\usage{
random_genotypes(genotypes = c("A", "H", "B"), size = 100, seed = NULL)
}
\arguments{
\item{genotypes}{Character vector containing the base genotypes.}

\item{size}{Output length.}

\item{seed}{Seed for reproducibility.}
}
\value{
Character vector containing the pseudo-random genotypes.
}
\description{
Generate a pseudo-random list of genotypes. Particularly useful for testing.
}
\examples{
MetaPipe::random_genotypes()
MetaPipe::random_genotypes(genotypes = c("nn", "np"))
MetaPipe::random_genotypes(size = 3)

}
\seealso{
\code{\link{random_map}}
}
