% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{random_map}
\alias{random_map}
\title{Generate a pseudo-random genetic map}
\usage{
random_map(
  genotypes = c("A", "H", "B"),
  lg = 1:10,
  markers = 10,
  population = 100,
  seed = NULL
)
}
\arguments{
\item{genotypes}{Character vector containing the base genotypes.}

\item{lg}{Numeric vector containing the linkage groups.}

\item{markers}{Number of markers per linkage group.}

\item{population}{Population size (rows).}

\item{seed}{Seed for reproducibility.}
}
\value{
Data frame containing the pseudo-random genetic map.
}
\description{
Generate a pseudo-random genetic map. Particularly useful for testing.
}
\examples{
gmap_1 <- MetaPipe::random_map()
gmap_2 <- MetaPipe::random_map(genotypes = c("nn", "np"))
gmap_3 <- MetaPipe::random_map(population = 3)

}
\seealso{
\code{\link{random_genotypes}}
}
