% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-qtl.R
\name{read.cross}
\alias{read.cross}
\title{Read QTL data}
\usage{
read.cross(geno, pheno, wdir = here::here(), ...)
}
\arguments{
\item{geno}{Data frame or string (filename) to file containing genotype 
data. For example a genetic map like \code{\link{father_riparia}}.}

\item{pheno}{Data frame or string (filename) to file containing phenotype 
data. For example, output from \code{\link{assess_normality}}.}

\item{wdir}{Working directory.}

\item{...}{Optional parameters for 
\code{\link[qtl::read.cross]{qtl:read.cross(...)}} function.}
}
\value{
Object of \code{cross} class for QTL mapping.
}
\description{
Read QTL data from two sources, one containing genotypes and another one
phenotypes.
}
\examples{
data(father_riparia)
data(ionomics)
ionomics_rev <- MetaPipe::replace_missing(ionomics, 
                                          excluded_columns = c(1, 2),
                                          replace_na =  TRUE)
ionomics_normalised <- 
  MetaPipe::assess_normality(ionomics_rev,
                             excluded_columns = c(1, 2),
                             out_prefix = "ionomics",
                             transf_vals = c(2, exp(1)),
                             show_stats = FALSE)
x_data <- MetaPipe::read.cross(father_riparia, 
                               ionomics_normalised$norm,
                               genotypes = c("nn", "np", "--"))

}
\seealso{
Other QTL mapping functions: 
\code{\link{qtl_perm_test}()},
\code{\link{qtl_scone}()}
}
\concept{QTL mapping functions}
