% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{replace_missing}
\alias{replace_missing}
\title{Replace missing values (NAs)}
\usage{
replace_missing(
  raw_data,
  excluded_columns,
  out_prefix = getwd(),
  prop_na = 0.5,
  replace_na = FALSE
)
}
\arguments{
\item{raw_data}{data frame containing the raw data}

\item{excluded_columns}{vector containing the indices of the data set properties, excluded columns}

\item{out_prefix}{prefix for output files, plots, etc.}

\item{prop_na}{proportion of missing values, if a feature exceeds this threshold, then it is dropped out}

\item{replace_na}{boolean flag to indicate whether or not missing values should be replaced by half of the minimum value}
}
\value{
data frame containing the raw data without missing values
}
\description{
Replace missing values (NAs)
}
\examples{
#' example_data <- data.frame(ID = c(1,2,3,4,5), 
                           P1 = c("one", "two", "three", "four", "five"), 
                           F1 = rnorm(5), 
                           F2 = rnorm(5))
example_data$F1[2:3] <- NA 
example_data$F2[4] <- NA 
replace_missing(example_data, c(2))
replace_missing(example_data, c(2), prop_na =  0.25)
replace_missing(example_data, c(2), replace_na =  TRUE)
}
