% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rplc_na}
\alias{rplc_na}
\title{Replace NAs}
\usage{
rplc_na(x)
}
\arguments{
\item{x}{original data points containing missing values (\code{NA}s)}
}
\value{
new data without missing entries, unless all of the original data
points were \code{NA}
}
\description{
Replace missing values (\code{NA}s) by half of the minimum value. If all the 
data points are missing, then returns the original input
}
\examples{
MetaPipe:::rplc_na(NA)
MetaPipe:::rplc_na(c(1, NA, 3, NA))

}
\keyword{internal}
