% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{save_plot}
\alias{save_plot}
\title{Save plot as PNG}
\usage{
save_plot(plt_obj, name, width = 6, height = 6)
}
\arguments{
\item{plt_obj}{Graphical object.}

\item{name}{Output name with or without path.}

\item{width}{Width in inches.}

\item{height}{Height in inches.}
}
\description{
Save a graphical object to disk in Portable Network Graphics (PNG) format.
}
\examples{
MetaPipe:::save_plot(hist(rnorm(100), 
                     main = "Histogram of Normal Distribution"), 
                     "hist")

# Clean up example outputs
MetaPipe:::tidy_up("hist")

}
\keyword{internal}
