% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{save_plotTIFF}
\alias{save_plotTIFF}
\title{Save plot as TIFF}
\usage{
save_plotTIFF(plt_obj, name, width = 6, height = 6)
}
\arguments{
\item{plt_obj}{Graphical object.}

\item{name}{Output name with or without path.}

\item{width}{Width in inches.}

\item{height}{Height in inches.}
}
\description{
Save a graphical object to disk in Tagged Image File Format (TIFF) format.
}
\examples{
MetaPipe::save_plotTIFF(hist(rnorm(100), 
                        main = "Histogram of Normal Distribution"), 
                        "hist")

}
\seealso{
\code{\link{save_plotPDF}} and \code{\link{save_plot}}
}
