% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.utils.R
\name{savePlotPDF}
\alias{savePlotPDF}
\title{Function to save a graphical object to disk in 
Portable Document Format (PDF) format.}
\usage{
savePlotPDF(myPlot, name, width = 6, height = 6)
}
\arguments{
\item{myPlot}{graphical object}

\item{name}{output name with or without path}

\item{width}{width in inches}

\item{height}{height in inches}
}
\description{
Function to save a graphical object to disk in 
Portable Document Format (PDF) format.
}
\examples{
norm_histo <- hist(rnorm(100), main = "Histogram of Normal Distribution")
savePlotPDF(norm_histo, "hist")

}
\seealso{
\code{\link{savePlot}} and \code{\link{savePlotTIFF}}
}
