% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_up}
\alias{tidy_up}
\title{List files matching a pattern and delete them}
\usage{
tidy_up(patterns, path = here::here())
}
\arguments{
\item{patterns}{Vector of strings with patterns to be deleted.}

\item{path}{Path to the directory containing the files}
}
\description{
List files matching a pattern and delete them
}
\keyword{internal}
