% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_data}
\alias{transform_data}
\title{Normalise data}
\usage{
transform_data(
  data,
  trait = "DATA",
  alpha = 0.05,
  index = NULL,
  transf_vals = c(2, exp(1), 3, 4, 5, 6, 7, 8, 9, 10),
  plots_prefix = file.path(tempdir(), "HIST"),
  digits = 6
)
}
\arguments{
\item{data}{Original data.}

\item{trait}{Trait name.}

\item{alpha}{Significance level.}

\item{index}{Index of the current trait.}

\item{transf_vals}{Transformation values.}

\item{plots_prefix}{Prefix for plots with or without path.}

\item{digits}{Significant digits to compare p-values of transformations.}
}
\value{
Transformed record.
}
\description{
Normalise data using different methods: log, power, and root, with a 
different number of parameters to find out which one transforms the data
into a normal-ish set.
}
\examples{
set.seed(123)
data <- rnorm(100, 5)
out_exp2 <- MetaPipe::transform_data(2 ^ data, "EXP_2")
out_root2 <- MetaPipe::transform_data(sqrt(data), "ROOT_2")
out_pow2 <- MetaPipe::transform_data(data ^ 2, "POW_2")

knitr::kable(head(out_exp2))
knitr::kable(head(out_root2))
knitr::kable(head(out_pow2))
}
