% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-qtl.R
\name{transform_pseudo_marker}
\alias{transform_pseudo_marker}
\title{Transform pseudo-marker}
\usage{
transform_pseudo_marker(x_data, marker, chr, pos)
}
\arguments{
\item{x_data}{Cross-data frame containing genetic map data and traits.}

\item{marker}{Pseudo-marker to be transformed.}

\item{chr}{Pseudo-marker's chromosome.}

\item{pos}{Pseudo-marker's position.}
}
\value{
A prime marker and its position.
}
\description{
Transform a pseudo-marker to a standard marker by looking for the nearest 
markers in a genetic map.
}
\examples{
# Toy dataset
excluded_columns <- c(1, 2)
population <- 5
seed <- 1
example_data <- data.frame(ID = 1:population,
                           P1 = c("one", "two", "three", "four", "five"),
                           T1 = rnorm(population),
                           T2 = rnorm(population))
example_data_normalised <- 
  data.frame(index = rep(c(1, 2), each = 5),
             trait = rep(c("T1", "T2"), each = 5),
             values = c(example_data$T1, example_data$T2),
             flag = "Normal",
             transf = "",
             transf_val = NA,
             stringsAsFactors = FALSE)

out_prefix <- here::here("metapipe")
example_data_normalised_post <- 
  MetaPipe:::assess_normality_postprocessing(example_data, 
                                            excluded_columns, 
                                            example_data_normalised,
                                            out_prefix = out_prefix)

# Create and store random genetic map (for testing only)
genetic_map <- 
  MetaPipe:::random_map(population = population, seed = seed)
write.csv(genetic_map, 
          here::here("metapipe_genetic_map.csv"), 
          row.names = FALSE)
# Load cross file with genetic map and raw data for normal traits
x <- qtl::read.cross(format = "csvs", 
                     dir = here::here(),
                     genfile = "metapipe_genetic_map.csv",
                     phefile = "metapipe_raw_data_norm.csv")
set.seed(seed)
x <- qtl::jittermap(x)
x <- qtl::calc.genoprob(x, step = 1, error.prob = 0.001)
MetaPipe:::transform_pseudo_marker(x, 'loc1', 1, 2.0)

}
\keyword{internal}
