% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{assess_normality_postprocessing}
\alias{assess_normality_postprocessing}
\title{Postprocessing for the normality assessmment of the features}
\usage{
assess_normality_postprocessing(
  raw_data,
  excluded_columns,
  raw_data_normalised,
  out_prefix = "metapipe",
  pareto_scaling = FALSE
)
}
\arguments{
\item{raw_data}{data frame containing the raw data}

\item{excluded_columns}{vector containing the indices of the data set properties, excluded columns}

\item{raw_data_normalised}{data frame containing the normalised raw data, created with \code{\link{assess_normality}}}

\item{out_prefix}{prefix for output files, plots, etc.}

\item{pareto_scaling}{booolean flag to indicate whether or not perform a pareto scaling on the normalised data}
}
\value{
Files containing the normal and non-paramatric normalised data are stored on disk
}
\description{
Postprocessing for the normality assessmment of the features
}
\examples{
example_data <- data.frame(ID = c(1,2,3,4,5), 
                           P1 = c("one", "two", "three", "four", "five"), 
                           F1 = rnorm(5), 
                           F2 = rnorm(5))
example_data_normalised <- assess_normality(example_data, c(1, 2))
assess_normality_postprocessing(example_data, c(1, 2), example_data_normalised)

}
\seealso{
\code{\link{assess_normality}} and \code{\link{assess_normality_stats}}
}
