% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{assess_normality_stats}
\alias{assess_normality_stats}
\title{Statistics for the normality assessment of the traits.}
\usage{
assess_normality_stats(out_prefix = "metapipe")
}
\arguments{
\item{out_prefix}{prefix for output files and plots}
}
\description{
This function uses an output file generated by \code{\link{assess_normality_postprocessing}}, 
therefore, it cannot be executed without previously running the postprocessing function.
}
\examples{
    \dontrun{
    example_data <- data.frame(ID = c(1,2,3,4,5), 
                               P1 = c("one", "two", "three", "four", "five"), 
                               T1 = rnorm(5), 
                               T2 = rnorm(5))
    example_data_normalised <- assess_normality(example_data, c(1, 2))
    assess_normality_postprocessing(example_data, c(1, 2), example_data_normalised)
    assess_normality_stats()
}

}
\seealso{
\code{\link{assess_normality}} and \code{\link{assess_normality_postprocessing}}
}
