% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{check_alpha}
\alias{check_alpha}
\title{Check significance level (alpha) validity}
\usage{
check_alpha(alpha)
}
\arguments{
\item{alpha}{significance level}
}
\value{
Nothing if alpha is valid, otherwise stop execution
}
\description{
Check significance level (alpha) validity
}
\examples{
check_alpha(0.5)
\dontrun{
    check_alpha(-0.5)
    check_alpha(1.5)
    check_alpha("1.5")
}
}
