% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-util.R
\name{compare_hist}
\alias{compare_hist}
\title{Title}
\usage{
compare_hist(original, transformed, feature, prefix, xlab)
}
\arguments{
\item{original}{Original data}

\item{transformed}{Transformed data}

\item{feature}{Feature name}

\item{prefix}{File prefix}

\item{xlab}{x-axis label}
}
\description{
Title
}
\examples{
norm_dist <- rnorm(100)
norm_dist_transformed <- norm_dist^2
compare_hist(norm_dist, norm_dist_transformed, "XYZ", "xyz_hist", "x")
}
