% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{father_riparia}
\alias{father_riparia}
\title{Father Riparia Genetic Map dataset}
\format{
A data frame with 296 rows and 1116 variables:
\describe{
    \item{ID}{Sample ID}
    \item{S<chr>_<pos> [1115 variables]}{Chromosome <chr>, position <pos>
    \itemize{
        \item \strong{S1_61235:}  Chromosome 1, position  61235
        \item \strong{S1_929902:} Chromosome 1, position 929902
        \item \strong{S1_666584:} Chromosome 1, position 666584
   }}
}
}
\source{
\url{https://fennell-lab.netlify.app}
}
\usage{
data(father_riparia)
}
\description{
A dataset containing information of a genetic map for a \emph{vitis Riparia} 
subject; more specifically, each column after \code{ID}, contains genotypes 
within each chromosome at different positions.
}
\author{
Gaurab Bhattarai \email{gbhattarai@uga.edu}
}
\keyword{datasets}
