% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_pipeline.R
\name{full_pipeline}
\alias{full_pipeline}
\title{Full Pipeline Function}
\usage{
full_pipeline(
  data,
  Project = NULL,
  Replication = NULL,
  Lab = NULL,
  DV = NULL,
  Group = NULL,
  output_path,
  folder_name = NULL,
  suppress_list_output = FALSE,
  method = "REML"
)
}
\arguments{
\item{data}{A list of data frames that contain the individual participant data. The function expects the relevant columns to be named consistently across all list objects. Relevant to this function are columns that represent information on the project (e.g., Many Labs 2), the replication (e.g., Ross1), the lab (the source a data point is assigned to), the group (either the treatment or control condition) and the single data point of the dependent variable per person.}

\item{Project}{Character vector with the name of the columns in the list elements of "data" that contain the project name(s). If \emph{is.null(Project) == TRUE}, "Project" is chosen as the default.}

\item{Replication}{Character vector with the name of the columns in the list elements of "data" that contain the replication name(s). If \emph{is.null(Replication) == TRUE}, "Replication" is chosen as the default. Each replication comprises a single effect with direct replications across multiple labs/sources.}

\item{Lab}{Character vector with the name of the columns in the list elements of "data" that contain the lab names. If \emph{is.null(Lab) == TRUE}, "Lab" is chosen as the default. The meta-analyses in MetaPipe::meta_analyses() and MetaPipe::full_pipeline() are run as random effects models in metafor::rma.mv() with “random = ~ 1 | Lab”. Thus, the pipeline assumes a distribution of true statistics (e.g., treatment means, mean differences, standardized mean differences).}

\item{DV}{Character vector with the name of the columns in the list elements of "data" that contain the (aggregated) dependent variable. If \emph{is.null(DV) == TRUE}, "DV" is chosen as the default.}

\item{Group}{Character vector with the name of the columns in the list elements of "data" that contain the (treatment/control) group identification. If \emph{is.null(Group) == TRUE}, "Group" is chosen as the default. These should only contain values of 0 (control group), 1 (treatment group) and NA (unidentified).}

\item{output_path}{Specify the output path for the full documentation of the MetaPipe pipeline. For an example of the exported structure please refer to https://github.com/JensFuenderich/MetaPipe/tree/main/Supplementary\%20Material/Table\%20Templates. If no folder is specified, the function will return its output only to the R environment (unless this is suppressed under suppress_list_output).}

\item{folder_name}{Optional character string to assign a custom name to the output folder. When folder_name is not specified, the folder name is set to “MetaPipe Output”.}

\item{suppress_list_output}{Logical. FALSE by default. If FALSE, the function will return a list output to the environment, containing the lab summaries and the codebook. If TRUE, these are not returned to the environment.}

\item{method}{Optional argument to specify the estimation method of the meta-analyses (the default is “REML”). For more information, please refer to the documentation of the metafor package.}
}
\value{
The output is a nested list object that represents the folder structure that is available under https://github.com/JensFuenderich/MetaPipe/tree/main/Supplementary\%20Material/Table\%20Templates
\subsection{Example}{

For an example, please refer to the github repository:
https://github.com/JensFuenderich/MetaPipe/blob/main/Supplementary\%20Material/Code\%20Examples/full_pipeline().R
}
}
\description{
\loadmathjax{}
\(\let\underscore_\)
}
\details{
The MetaPipe pipeline is a tool to provide structure to the meta-analytical-analyses of multi-lab replication projects. A flowchart that depicts the whole process is available at https://github.com/JensFuenderich/MetaPipe/blob/main/Supplementary\%20Material/MetaPipe\%20Flow\%20Chart.pdf
The dark blue blocks are .csv files. The dark green blocks each refer to a step in the pipeline that is performed by a MetaPipe function. MetaPipe::full_pipeline() performs all of these steps and returns "MetaPipe_Data.csv" which may be provided to the MetaPipe App for handy data selection and basic plotting of the analysis results.
For an example of the MetaPipe Output structure, please refer to https://github.com/JensFuenderich/MetaPipe/tree/main/Supplementary\%20Material/Table\%20Templates
}
