% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{generate_hist}
\alias{generate_hist}
\title{Generate histogram}
\usage{
generate_hist(
  data,
  title,
  prefix = "metapipe",
  xlab = NULL,
  save = TRUE,
  alpha = 1,
  angle = 60,
  bins = 20,
  col = "black",
  hjust = 1,
  fill = "#7FCDBB",
  is_trait = FALSE
)
}
\arguments{
\item{data}{Histogram data.}

\item{title}{Plot title.}

\item{prefix}{File prefix.}

\item{xlab}{X-axis label.}

\item{save}{Boolean flag to indicate if that plot should be save to disk.}

\item{alpha}{Numeric value to set the transparency of the bins.}

\item{angle}{Rotation angle for labels.}

\item{bins}{Number of bins to plot.}

\item{col}{Colour for bins border.}

\item{hjust}{Horizontal adjustment of labels.}

\item{fill}{Filling colour.}

\item{is_trait}{Boolean flag to prepend "Trait" to \code{title}.}
}
\description{
Generate histogram
}
\examples{
norm_dist <- rnorm(100)
MetaPipe::generate_hist(norm_dist, "XYZ", "xyz_hist", "x")
MetaPipe::generate_hist(norm_dist, "XYZ", "xyz_hist", "x", is_trait = TRUE)
}
