% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{generate_hist}
\alias{generate_hist}
\title{This function generates a single histogram}
\usage{
generate_hist(
  data,
  title,
  prefix = "metapipe",
  xlab = NULL,
  save = TRUE,
  alpha = 1,
  angle = 60,
  bins = 20,
  col = "black",
  hjust = 1,
  fill = "#7FCDBB",
  is_trait = FALSE
)
}
\arguments{
\item{data}{histogram data}

\item{title}{plot title}

\item{prefix}{file prefix [default: \code{metapipe}]}

\item{xlab}{x-axis label [default: \code{NULL}]}

\item{save}{boolean flag to indicate if that plot should be save to disk [default: \code{TRUE}]}

\item{alpha}{value to set the transparency of the bins [default: \code{1}]}

\item{angle}{rotation angle for labels [default: \code{60}]}

\item{bins}{number of bins to plot [default: \code{20}]}

\item{col}{colour for bins border [default: \code{"black"}]}

\item{hjust}{horizontal adjustment of labels [default: \code{1}]}

\item{fill}{filling colour [default: \code{"#7FCDBB"}]}

\item{is_trait}{if TRUE then prepends "Trait" to title [default: \code{FALSE}]}
}
\description{
This function generates a single histogram
}
\examples{
norm_dist <- rnorm(100)
generate_hist(norm_dist, "XYZ", "xyz_hist", "x")
generate_hist(norm_dist, "XYZ", "xyz_hist", "x", is_trait = TRUE)
}
