% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{generate_hist}
\alias{generate_hist}
\title{This function generates a single histogram}
\usage{
generate_hist(data, title, prefix = "metapipe", xlab = NULL, save = TRUE)
}
\arguments{
\item{data}{histogram data}

\item{title}{plot title}

\item{prefix}{file prefix [default: metapipe]}

\item{xlab}{x-axis label [default: NULL]}

\item{save}{boolean flag to indicate if that plot should be save to disk [default: TRUE]}
}
\description{
This function generates a single histogram
}
\examples{
norm_dist <- rnorm(100)
generate_hist(norm_dist, "XYZ", "xyz_hist", "x")
}
