% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-util.R
\name{ggplot_save}
\alias{ggplot_save}
\title{Function to save a graphical object generated with GGPLOT2 to disk in 
Portable Network Graphics (PNG) format.}
\usage{
ggplot_save(myPlot, name, width = 6, height = 6)
}
\arguments{
\item{myPlot}{graphical object}

\item{name}{output name with or without path}

\item{width}{width in inches}

\item{height}{height in inches}
}
\description{
Function to save a graphical object generated with GGPLOT2 to disk in 
Portable Network Graphics (PNG) format.
}
\examples{
myplot <- ggplot2::qplot(rnorm(100))
ggplot_save(myplot, "hist")
}
