% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{hex_logo}
\alias{hex_logo}
\title{Create Hexagonal logo for MetaPipe}
\usage{
hex_logo(
  subplot = system.file("images/lab-2.png", package = "MetaPipe"),
  dpi = 800,
  h_color = "#000000",
  h_fill = "#363b74",
  output = system.file("images/metapipe.png", package = "MetaPipe"),
  p_color = "#eeeeee"
)
}
\arguments{
\item{subplot}{subplot/image/icon to be used in the background}

\item{dpi}{dots per inch}

\item{h_color}{hexagon edge colour}

\item{h_fill}{hexagon filling colour}

\item{output}{output filename and path}

\item{p_color}{package name colour}
}
\description{
Create Hexagonal logo for MetaPipe
}
\examples{
\dontrun{
    hex_logo()
}
}
