% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl-utils.R
\name{is_pseudo_marker}
\alias{is_pseudo_marker}
\title{Detect pseudo-markers}
\usage{
is_pseudo_marker(marker)
}
\arguments{
\item{marker}{String with the marker.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Verifies whether or not a string is a pseudo-marker, these contain the 
substring \code{'loc'}.
}
\examples{
is_pseudo_marker('c1.loc1')
is_pseudo_marker('S1_2345')
}
