% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-qtl.R
\name{is_pseudo_marker}
\alias{is_pseudo_marker}
\title{Detect pseudo-markers}
\usage{
is_pseudo_marker(marker)
}
\arguments{
\item{marker}{String with the marker.}
}
\value{
\code{TRUE} if the given marker is a pseudo-marker, \code{FALSE} 
otherwise.
}
\description{
Verifies whether or not a string is a pseudo-marker, these contain the 
substring \code{'loc'}.
}
\examples{
MetaPipe:::is_pseudo_marker('c1.loc1')
MetaPipe:::is_pseudo_marker('S1_2345')

}
\keyword{internal}
