% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{load_raw}
\alias{load_raw}
\title{Load raw data}
\usage{
load_raw(raw_data_filename, excluded_columns)
}
\arguments{
\item{raw_data_filename}{filename containing the raw data, with or without full path}

\item{excluded_columns}{vector containing the indices of the data set properties, excluded columns}
}
\value{
pre-processed raw data [data frame]
}
\description{
Load raw data
}
\examples{
\dontrun{
    example_data <- data.frame(ID = c(1,2,3,4,5), 
                               P1 = c("one", "two", "three", "four", "five"), 
                               T1 = rnorm(5), 
                               T2 = rnorm(5))
    write.csv(example_data, "example_data.csv", row.names = FALSE)
    write.csv(example_data[c(1:5, 1, 2), ], "example_data_dup.csv", row.names = FALSE)
    load_raw("example_data.csv", c(1, 2))
    load_raw("example_data_dup.csv", c(1, 2))
}
}
