% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{log_transformation}
\alias{log_transformation}
\title{Normalise data with a log transformation}
\usage{
log_transformation(
  data,
  feature,
  alpha = 0.05,
  transf = c(2, exp(1), 3, 4, 5, 6, 7, 8, 9, 10),
  plots_prefix = "HIST",
  digits = 6
)
}
\arguments{
\item{data}{original data}

\item{feature}{feature name}

\item{alpha}{significance level}

\item{transf}{transformation values}

\item{plots_prefix}{prefix for plots with or without path}

\item{digits}{significant digits to compare p-values of transformations}
}
\value{
data structure containing the normalised data, NULL if no 
transformation was performed
}
\description{
Normalise data with a log transformation
}
\examples{
set.seed(123)
data <- rnorm(100)
log_transformation(2^data, "POW_2")
}
