% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{paretoscale}
\alias{paretoscale}
\title{Pareto scaling function
This function is adapted from Stephen C. Grace and Dane A. Hudson}
\usage{
paretoscale(z)
}
\arguments{
\item{z}{original 2-dimensional data}
}
\value{
normalised data
}
\description{
Pareto scaling function
This function is adapted from Stephen C. Grace and Dane A. Hudson
}
\examples{
set.seed(123)
data <- matrix(rnorm(100, 5), ncol = 2)
data_new <- paretoscale(data)
}
