% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{paretoscale}
\alias{paretoscale}
\title{Pareto scaling}
\usage{
paretoscale(z)
}
\arguments{
\item{z}{Original 2-dimensional data.}
}
\value{
Scaled data.
}
\description{
Perform a pareto scaling.
This function is adapted from Stephen C. Grace and Dane A. Hudson.
}
\examples{
set.seed(123)
data <- matrix(rnorm(100, 5), ncol = 2)
data_new <- MetaPipe:::paretoscale(data)

}
\keyword{internal}
