% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{PCA}
\alias{PCA}
\title{Simple PCA}
\usage{
PCA(data, name = "PCA", plot = TRUE, width = 6, height = 6, ...)
}
\arguments{
\item{data}{A numeric or complex matrix (or data frame) that will be used to
perform the Principal Components Analysis.}

\item{name}{Output name with or without path.}

\item{plot}{Boolean flag to indicate whether or not to create a PCA biplot.}

\item{width}{Width in inches.}

\item{height}{Height in inches.}

\item{...}{Arguments passed on to 
\code{\link[factoextra:fviz_pca_biplot]{factoextra::fviz_pca_biplot}}.}
}
\value{
Data frame with PCA result.
}
\description{
Perform a simple PCA using \code{\link[stats:prcomp]{stats::prcomp}}. 
Optionally, it will create a PCA biplot using 
\code{\link[factoextra:fviz_pca_biplot]{factoextra::fviz_pca_biplot}} if 
\code{plot = TRUE}.
}
\examples{
\donttest{
out_prefix <- file.path(tempdir(), "metapipe_PCA")
# Toy dataset
example_data <- data.frame(ID = c(1,2,3,4,5), 
                           P1 = c("one", "two", "three", "four", "five"), 
                           T1 = rnorm(5), 
                           T2 = rnorm(5))
example_data_pca <- MetaPipe::PCA(example_data[, -c(1:2)],
                                  name = out_prefix)

# F1 Seedling Ionomics dataset
data(ionomics) # Includes some missing data
ionomics_rev <- MetaPipe::replace_missing(ionomics, 
                                          excluded_columns = c(1, 2),
                                          replace_na =  TRUE,
                                          out_prefix = out_prefix)
ionomics_pca <- MetaPipe::PCA(ionomics_rev[, -c(1:2)],
                              name = out_prefix)
}
}
