% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{power_transformation}
\alias{power_transformation}
\title{Power transformation}
\usage{
power_transformation(
  data,
  trait = "DATA",
  alpha = 0.05,
  transf_vals = c(2, exp(1), 3, 4, 5, 6, 7, 8, 9, 10),
  plots_prefix = file.path(tempdir(), "HIST"),
  digits = 6,
  plot = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{data}{Original data.}

\item{trait}{Trait name.}

\item{alpha}{Significance level.}

\item{transf_vals}{Transformation values.}

\item{plots_prefix}{Prefix for plots with or without path.}

\item{digits}{Significant digits to compare p-values of transformations.}

\item{plot}{Boolean flag to indicate whether or not to plot a histogram 
comparing the original and transformed/normalised data.}

\item{quiet}{Boolean flag to hide status messages.}
}
\value{
Data structure containing the normalised data, original data if no 
transformation was performed.
}
\description{
Normalise data with a power transformation.
}
\examples{
set.seed(123)
data <- rnorm(100, 5)
transformed <- MetaPipe::power_transformation(sqrt(data), "ROOT_2")

}
\seealso{
Other Transformation functions: 
\code{\link{log_transformation}()},
\code{\link{root_transformation}()}
}
\concept{Transformation functions}
