% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{qtl_perm_test}
\alias{qtl_perm_test}
\title{Perform QTL mapping permutation test using \code{scanone} to obtain significant QTLs}
\usage{
qtl_perm_test(
  x_data,
  cpus = 1,
  qtl_method = "par-scanone",
  raw_data_normalised = NULL,
  lod_threshold = 3,
  parametric = TRUE,
  n_perm = 1000,
  plots_dir = getwd(),
  ...
)
}
\arguments{
\item{x_data}{cross-data containing genetic map data and traits}

\item{cpus}{number of CPUS to be used}

\item{qtl_method}{QTL mapping method [default: \code{scanone}]}

\item{raw_data_normalised}{normalised raw data, see \code{\link{assess_normality}}}

\item{lod_threshold}{LOD score threshold to look up for significant QTLs}

\item{parametric}{boolean flag to indicate whether or not \code{x_data} is parametric}

\item{n_perm}{number of permutations}

\item{plots_dir}{output directory for plots}

\item{...}{optional parameters for \code{R/qtl} library}
}
\value{
data frame containing the significant QTLs information
}
\description{
Perform QTL mapping permutation test using \code{scanone} to obtain significant QTLs
}
\examples{
\dontrun{
    # Create toy dataset
    excluded_columns <- c(2)
    population <- 5
    seed <- 1
    example_data <- data.frame(ID = 1:population,
                               P1 = c("one", "two", "three", "four", "five"),
                               T1 = rnorm(population),
                               T2 = rnorm(population))
    example_data_normalised <- data.frame(index = rep(c(1, 2), each = 5),
                                          trait = rep(c("T1", "T2"), each = 5),
                                          values = c(example_data$T1, example_data$T2),
                                          flag = "Normal",
                                          transf = "",
                                          transf_val = NA,
                                          stringsAsFactors = FALSE)
    assess_normality_postprocessing(example_data, excluded_columns, 
                                    example_data_normalised, 
                                    out_prefix = here::here("metapipe"))
    
    # Create and store random genetic map [for testing only]
    genetic_map <- random_map(population = population, seed = seed)
    write.csv(genetic_map, here::here("metapipe_genetic_map.csv"), row.names = FALSE)
    
    # Load cross file with genetic map and raw data for normal traits
    x <- qtl::read.cross(format = "csvs", 
                         dir = here::here(),
                         genfile = "metapipe_genetic_map.csv",
                         phefile = "metapipe_raw_data_norm.csv")
    set.seed(seed)
    x <- qtl::jittermap(x)
    x <- qtl::calc.genoprob(x, step = 1, error.prob = 0.001)
    x_qtl_perm <- qtl_perm_test(x, n_perm = 5, model = "normal", method = "hk")
}

}
\seealso{
\code{\link{assess_normality}} and \code{\link{qtl_scone}}
}
