% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{qtl_perm_test}
\alias{qtl_perm_test}
\title{QTL mapping permutation test}
\usage{
qtl_perm_test(
  x_data,
  cpus = 1,
  qtl_method = "par-scanone",
  raw_data_normalised = NULL,
  lod_threshold = 3,
  parametric = TRUE,
  n_perm = 1000,
  plots_dir = getwd(),
  ...
)
}
\arguments{
\item{x_data}{Cross-data frame containing genetic map data and traits.}

\item{cpus}{Number of CPUs to be used in the computation.}

\item{qtl_method}{QTL mapping method.}

\item{raw_data_normalised}{Normalised raw data, see 
\code{\link{assess_normality}}.}

\item{lod_threshold}{LOD score threshold to look up for significant QTLs}

\item{parametric}{Boolean flag to indicate whether or not \code{x_data} 
contains parametric (normal) traits.}

\item{n_perm}{Number of permutations.}

\item{plots_dir}{Output directory for plots.}

\item{...}{Optional parameters for 
\code{\link[qtl:scanone]{qtl:scanone(...)}}.}
}
\value{
Data frame containing the significant QTLs information.
}
\description{
Perform QTL mapping permutation test using the 
\code{\link[qtl:scanone]{qtl:scanone(...)}} function to find significant QTL.
}
\examples{
\dontrun{
    # Create toy dataset
    excluded_columns <- c(2)
    population <- 5
    seed <- 1
    example_data <- data.frame(ID = 1:population,
                               P1 = c("one", "two", "three", "four", "five"),
                               T1 = rnorm(population),
                               T2 = rnorm(population))
    example_data_normalised <- data.frame(index = rep(c(1, 2), each = 5),
                                          trait = rep(c("T1", "T2"), each = 5),
                                          values = c(example_data$T1, example_data$T2),
                                          flag = "Normal",
                                          transf = "",
                                          transf_val = NA,
                                          stringsAsFactors = FALSE)
    assess_normality_postprocessing(example_data, excluded_columns, 
                                    example_data_normalised, 
                                    out_prefix = here::here("metapipe"))
    
    # Create and store random genetic map [for testing only]
    genetic_map <- random_map(population = population, seed = seed)
    write.csv(genetic_map, here::here("metapipe_genetic_map.csv"), row.names = FALSE)
    
    # Load cross file with genetic map and raw data for normal traits
    x <- qtl::read.cross(format = "csvs", 
                         dir = here::here(),
                         genfile = "metapipe_genetic_map.csv",
                         phefile = "metapipe_raw_data_norm.csv")
    set.seed(seed)
    x <- qtl::jittermap(x)
    x <- qtl::calc.genoprob(x, step = 1, error.prob = 0.001)
    x_qtl_perm <- qtl_perm_test(x, n_perm = 5, model = "normal", method = "hk")
}

}
\seealso{
\code{\link{assess_normality}} and \code{\link{qtl_scone}}
}
