% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{qtl_scone}
\alias{qtl_scone}
\title{Perform QTL mapping scanone to obtain LOD scores for all features and markers}
\usage{
qtl_scone(x_data, features, cpus = 1, ...)
}
\arguments{
\item{x_data}{cross-file containing genetic map data and features}

\item{features}{feature names}

\item{cpus}{number of CPUS to be used}

\item{...}{S4 parameters for R/qtl library}
}
\value{
data frame containing the LOD scores
}
\description{
Perform QTL mapping scanone to obtain LOD scores for all features and markers
}
