% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{qtl_scone}
\alias{qtl_scone}
\title{Perform QTL mapping scanone to obtain LOD scores for all features and markers}
\usage{
qtl_scone(x_data, cpus = 1, ...)
}
\arguments{
\item{x_data}{cross-file containing genetic map data and features}

\item{cpus}{number of CPUS to be used}

\item{...}{S4 parameters for R/qtl library}
}
\value{
data frame containing the LOD scores
}
\description{
Perform QTL mapping scanone to obtain LOD scores for all features and markers
}
\examples{
# Create toy dataset
excluded_columns <- c(2)
population <- 5
seed <- 1
example_data <- data.frame(ID = 1:population,
                           P1 = c("one", "two", "three", "four", "five"),
                           F1 = rnorm(population),
                           F2 = rnorm(population))
example_data_normalised <- data.frame(index = rep(c(1, 2), each = 5),
                                      feature = rep(c("F1", "F2"), each = 5),
                                      values = c(example_data$F1, example_data$F2),
                                      flag = "Normal",
                                      transf = "",
                                      transf.value = NA,
                                      stringsAsFactors = FALSE)
assess_normality_postprocessing(example_data, excluded_columns, example_data_normalised)

# Create and store random genetic map [for testing only]
genetic_map <- random_map(population = population, seed = seed)
write.csv(genetic_map, here::here("metapipe_genetic_map.csv"), row.names = FALSE)
print(list.files(here::here()))
# Load cross file with genetic map and raw data for normal features
x <- qtl::read.cross(format = "csvs", 
                     dir = here::here(),
                     genfile = "metapipe_genetic_map.csv",
                     phefile = "metapipe_raw_data_norm.csv")
set.seed(seed)
x <- qtl::jittermap(x)
x <- qtl::calc.genoprob(x, step = 1, error.prob = 0.001)
x_scone <- MetaPipe::qtl_scone(x, 1)
}
