% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{random_genotypes}
\alias{random_genotypes}
\title{Generate a pseudo-random list of genotypes. Particularly useful for testing.}
\usage{
random_genotypes(genotypes = c("A", "H", "B"), size = 100, seed = NULL)
}
\arguments{
\item{genotypes}{vector containing the base genotypes}

\item{size}{output length}

\item{seed}{seed for reproducibility}
}
\value{
vector containing the pseudo-random genotypes
}
\description{
Generate a pseudo-random list of genotypes. Particularly useful for testing.
}
\examples{
random_genotypes()
random_genotypes(genotypes = c("nn", "np"))
random_genotypes(size = 3)

}
\seealso{
\code{\link{random_map}}
}
