% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{random_map}
\alias{random_map}
\title{Generate a pseudo-random genetic map. Particularly useful for testing.}
\usage{
random_map(
  genotypes = c("A", "H", "B"),
  lg = 1:10,
  markers = 10,
  population = 100,
  seed = NULL
)
}
\arguments{
\item{genotypes}{vector containing the base genotypes}

\item{lg}{vector containing the linkage groups}

\item{markers}{number of markers per linkage group}

\item{population}{population size (rows)}

\item{seed}{seed for reproducibility}
}
\value{
data frame containing the pseudo-random genetic map.
}
\description{
Generate a pseudo-random genetic map. Particularly useful for testing.
}
\examples{
random_map()
random_map(genotypes = c("nn", "np"))
random_map(population = 3)
}
