% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapipe.R
\name{replace_missing}
\alias{replace_missing}
\title{Replace missing values (NAs)}
\usage{
replace_missing(
  raw_data,
  excluded_columns,
  out_prefix = "metapipe",
  prop_na = 0.5,
  replace_na = FALSE
)
}
\arguments{
\item{raw_data}{data frame containing the raw data}

\item{excluded_columns}{vector containing the indices of the data set properties, excluded columns}

\item{out_prefix}{prefix for output files, plots, etc.}

\item{prop_na}{proportion of missing values, if a feature exceeds this threshold, then it is dropped out}

\item{replace_na}{boolean flag to indicate whether or not missing values should be replaced by half of the minimum value}
}
\value{
data frame containing the raw data without missing values
}
\description{
Replace missing values (NAs)
}
\examples{
example_data <- data.frame(ID = c(1,2,3,4,5), 
                           P1 = c("one", "two", "three", "four", "five"), 
                           T1 = rnorm(5), 
                           T2 = rnorm(5),
                           T3 = c(NA, rnorm(4)),                  #  20 \% NAs
                           T4 = c(NA, 1.2, -0.5, NA, 0.87),       #  40 \% NAs
                           T5 = NA)                               # 100 \% NAs
replace_missing(example_data, c(1, 2))
replace_missing(example_data, c(1, 2), prop_na =  0.25)
replace_missing(example_data, c(1, 2), replace_na =  TRUE)
}
