% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{save_plot}
\alias{save_plot}
\title{Function to save a graphical object to disk in 
Portable Network Graphics (PNG) format.}
\usage{
save_plot(myPlot, name, width = 6, height = 6)
}
\arguments{
\item{myPlot}{graphical object}

\item{name}{output name with or without path}

\item{width}{width in inches}

\item{height}{height in inches}
}
\description{
Function to save a graphical object to disk in 
Portable Network Graphics (PNG) format.
}
\examples{
\dontrun{
    save_plot(hist(rnorm(100), 
              main = "Histogram of Normal Distribution"), 
              "hist")
}

}
\seealso{
\code{\link{save_plotPDF}} and \code{\link{save_plotTIFF}}
}
