% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{save_plotPDF}
\alias{save_plotPDF}
\title{Save plot as PDF}
\usage{
save_plotPDF(plt_obj, name, width = 6, height = 6)
}
\arguments{
\item{plt_obj}{Graphical object.}

\item{name}{Output name with or without path.}

\item{width}{Width in inches.}

\item{height}{Height in inches.}
}
\description{
Save a graphical object to disk in Portable Document Format (PDF) format.
}
\examples{
MetaPipe::save_plotPDF(hist(rnorm(100), 
                       main = "Histogram of Normal Distribution"), 
                       "hist")

}
\seealso{
\code{\link{save_plot}} and \code{\link{save_plotTIFF}}
}
