% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-util.R
\name{save_plotPDF}
\alias{save_plotPDF}
\title{Function to save a graphical object to disk in 
Portable Document Format (PDF) format.}
\usage{
save_plotPDF(myPlot, name, width = 6, height = 6)
}
\arguments{
\item{myPlot}{graphical object}

\item{name}{output name with or without path}

\item{width}{width in inches}

\item{height}{height in inches}
}
\description{
Function to save a graphical object to disk in 
Portable Document Format (PDF) format.
}
\examples{
save_plotPDF(hist(rnorm(100), main = "Histogram of Normal Distribution"), 
"hist")

}
\seealso{
\code{\link{save_plot}} and \code{\link{save_plotTIFF}}
}
