% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-util.R
\name{save_plotTIFF}
\alias{save_plotTIFF}
\title{Function to save a graphical object to disk in 
Tagged Image File Format (TIFF) format.}
\usage{
save_plotTIFF(myPlot, name, width = 6, height = 6)
}
\arguments{
\item{myPlot}{graphical object}

\item{name}{output name with or without path}

\item{width}{width in inches}

\item{height}{height in inches}
}
\description{
Function to save a graphical object to disk in 
Tagged Image File Format (TIFF) format.
}
\examples{
save_plotTIFF(hist(rnorm(100), main = "Histogram of Normal Distribution"), 
"hist")

}
\seealso{
\code{\link{save_plotPDF}} and \code{\link{save_plot}}
}
