% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_data}
\alias{transform_data}
\title{Normalise data using different methods: log, power, and root, with a 
different number of parameters to find out which one transforms the data
into a normal-ish set.}
\usage{
transform_data(
  data,
  feature = NULL,
  alpha = 0.05,
  index = NULL,
  transf_vals = c(2, exp(1), 3, 4, 5, 6, 7, 8, 9, 10),
  plots_prefix = "HIST",
  digits = 6
)
}
\arguments{
\item{data}{original data}

\item{feature}{feature name}

\item{alpha}{significance level}

\item{index}{index of the current feature}

\item{transf_vals}{transformation values}

\item{plots_prefix}{prefix for plots with or without path}

\item{digits}{significant digits to compare p-values of transformations}

\item{ref_pval}{reference p-value}
}
\value{

}
\description{
Normalise data using different methods: log, power, and root, with a 
different number of parameters to find out which one transforms the data
into a normal-ish set.
}
\examples{
set.seed(123)
data <- rnorm(100, 5)
transform_data(2 ^ data, "EXP_2")
transform_data(sqrt(data), "ROOT_2")
transform_data(data ^ 2, "POW_2")
}
